### 2021-01-26 3e10a5e
update the Calibrate Barcode window
