### 2021-01-27 #2980 @gkthiruvathukal
@yc015, thanks for the nice greeting! Can I ask you to add a couple more who are not on the editorial board or EICs?  I'm only asking that because our board in general is facing heavy workload now. (I personally am interested in this topic, since it has a nice overlap with Digital Humanities, one of my areas of interest.
