### 2021-02-20 #2980 @gkthiruvathukal
@stsievert and @thomakra, are you willing to contribute a review for this JOSS submission?

@yc015 Many apologies for the delay in following up on your suggestions!
