+++ kalmus/tkinter_windows/plot_barcodes_windows/PlotBarcodeWindow.py
@@ -14,9 +14,9 @@ from matplotlib.backends.backend_tkagg import (FigureCanvasTkAgg, NavigationTool
 import numpy as np
 import os
 
-from kalmus.visualization_utils import show_colors_in_cube
+from kalmus.utils.visualization_utils import show_colors_in_cube
 from kalmus.tkinter_windows.KALMUS_utils import resource_path, update_hist
-import kalmus.Artist
+import kalmus.utils.Artist
 from skimage.color import rgb2hsv
 
 
@@ -238,7 +238,7 @@ class OutputCSVWindow():
             # Data frame of the csv file for the color barcode
             data_frame = ['Frame index', 'Red (0-255)', 'Green (0-255)', 'Blue (0-255)', 'Hue (0 -360)',
                           'Saturation (0 - 1)', 'Value (lightness) (0 - 1)','Brightness']
-            kalmus.Artist.write_in_info(data_frame, csv_filename, mode='w')
+            kalmus.utils.Artist.write_in_info(data_frame, csv_filename, mode='w')
 
             cur_frame = starting_frame
             # Get the per frame level color data
@@ -252,19 +252,19 @@ class OutputCSVWindow():
                 s = float(hsv[1])
                 v = float(hsv[2])
 
-                kalmus.Artist.write_in_info([cur_frame, r, g, b, h, s, v, brightness], csv_filename)
+                kalmus.utils.Artist.write_in_info([cur_frame, r, g, b, h, s, v, brightness], csv_filename)
 
                 cur_frame += sample_rate
 
         elif self.barcode.barcode_type == 'Brightness':
             # Data frame of the csv file for the brightness barcode
             data_frame = ['Frame index', 'Brightness']
-            kalmus.Artist.write_in_info(data_frame, csv_filename, mode='w')
+            kalmus.utils.Artist.write_in_info(data_frame, csv_filename, mode='w')
 
             cur_frame = starting_frame
             # Get the per frame level brightness data
             for bri in self.barcode.brightness:
-                kalmus.Artist.write_in_info([cur_frame, int(bri[0])], csv_filename)
+                kalmus.utils.Artist.write_in_info([cur_frame, int(bri[0])], csv_filename)
 
                 cur_frame += sample_rate
 
