+++ kalmus/tkinter_windows/meta_info_windows/WhichBarcodeCheckMeta.py
@@ -1,7 +1,7 @@
 """ WhichBarcodeCheckMeta Class """
 
 import tkinter
-from kalmus.tkinter_windows.MetaInfoWindow import MetaInfoWindow
+from kalmus.tkinter_windows.meta_info_windows.MetaInfoWindow import MetaInfoWindow
 from kalmus.tkinter_windows.KALMUS_utils import resource_path
 
 
