+++ kalmus/tkinter_windows/meta_info_windows/MetaInfoWindow.py
@@ -1,7 +1,7 @@
 """ MetaInfoWindow Class """
 
 import tkinter
-from kalmus.tkinter_windows.SpecifyMetaDataWindow import SpecifyMetaDataWindow
+from kalmus.tkinter_windows.meta_info_windows.SpecifyMetaDataWindow import SpecifyMetaDataWindow
 from kalmus.tkinter_windows.KALMUS_utils import resource_path
 
 keys = ["Film Title", "Directors", "Country of Origin", "Produced Year", "Genre"]
