+++ kalmus/tkinter_windows/MainWindowVersion2.py
@@ -69,7 +69,7 @@ class MainWindow():
         self.root = tkinter.Tk()
 
         self.root.configure(bg='#85C1FA')
-        self.root.wm_title("KALMUS Version 1.3.5")
+        self.root.wm_title("KALMUS Version 1.3.6 alpha0")
         self.root.iconbitmap(resource_path("kalmus_icon.ico"))
 
         # Initialize the figure
