### 2021-02-21 #2980 @stsievert
Thanks for the request! I'm afraid I'll have to decline this review – my bandwidth is unfortunately limited.

I have briefly glanced at the code. I would recommend reviewing the JOSS reviewer guidelines and improving the documentation.
