+++ requirements.txt
@@ -1,96 +1,62 @@
 attrs==19.1.0
 backcall==0.1.0
-biopython==1.77
-bleach==3.1.0
-brotlipy==0.7.0
-certifi==2020.6.20
-cffi==1.14.0
-chardet==3.0.4
-cloudpickle==1.4.1
-colorama==0.4.3
-cryptography==2.9.2
+biopython==1.78
+bleach>=3.3.0
+certifi==2020.12.5
 cycler==0.10.0
-cytoolz==0.10.1
-dask==2.16.0
 decorator==4.4.0
 defusedxml==0.6.0
 entrypoints==0.3
-idna==2.9
-imageio==2.8.0
-importlib-metadata==1.5.0
+imageio==2.9.0
 ipykernel==5.1.1
-ipympl==0.5.6
 ipython==7.7.0
 ipython-genutils==0.2.0
 ipywidgets==7.5.0
 jedi==0.14.1
 Jinja2==2.10.1
-joblib==0.15.1
-json5==0.9.0
+joblib==1.0.0
 jsonschema==3.0.1
 jupyter==1.0.0
-jupyter-client==5.3.1
 jupyter-console==6.0.0
-jupyter-core==4.5.0
-jupyterlab==2.1.3
-jupyterlab-server==1.1.5
-kiwisolver==1.1.0
+kalmus==1.3.6
+kiwisolver==1.3.1
 MarkupSafe==1.1.1
-matplotlib==3.2.1
+matplotlib==3.2.2
 mistune==0.8.4
-mkl-fft==1.0.15
-mkl-random==1.1.0
-mkl-service==2.3.0
-mlxtend==0.17.2
 mpmath==1.1.0
 nbconvert==5.5.0
 nbformat==4.4.0
-networkx==2.4
+networkx==2.5
 nose==1.3.7
 notebook==6.0.0
-numpy==1.18.1
-olefile==0.46
-pandas==1.0.3
+numpy==1.19.4
+opencv-python==4.4.0.46
 pandocfilters==1.4.2
 parso==0.5.1
 pickleshare==0.7.5
-Pillow==7.1.2
+Pillow==8.0.1
 prometheus-client==0.7.1
 prompt-toolkit==2.0.9
-pycparser==2.20
 Pygments==2.4.2
-pyOpenSSL==19.1.0
 pyparsing==2.4.1.1
-PyQt5==5.12.3
-PyQt5-sip==4.19.18
-PyQtChart==5.12
-PyQtWebEngine==5.12.1
 pyrsistent==0.15.4
-PySocks==1.7.1
 python-dateutil==2.8.0
 pytz==2019.1
 PyWavelets==1.1.1
-pywin32==227
 pywinpty==0.5.5
-pyzmq==19.0.1
 qtconsole==4.5.2
-QtPy==1.9.0
-requests==2.23.0
 scikit-image==0.16.2
-scikit-learn==0.22.1
-scipy==1.4.1
+scikit-learn==0.24.0
+scipy==1.5.4
 Send2Trash==1.5.0
 six==1.12.0
 sympy==1.4
 terminado==0.8.2
 testpath==0.4.2
-toolz==0.10.0
-tornado==6.0.3
+threadpoolctl==2.1.0
+tifffile==2020.12.8
 traitlets==4.3.2
-urllib3==1.25.9
 wcwidth==0.1.7
 webencodings==0.5.1
 widgetsnbextension==3.5.0
-win-inet-pton==1.1.0
 wincertstore==0.2
-zipp==3.1.0
