+++ requirements.txt
@@ -28,7 +28,7 @@ nbconvert==5.5.0
 nbformat==4.4.0
 networkx==2.5
 nose==1.3.7
-notebook==6.0.0
+notebook>=6.1.5
 numpy==1.19.4
 opencv-python==4.4.0.46
 pandocfilters==1.4.2
