### 2021-02-21 #2980 @gkthiruvathukal
@stsievert I know the feeling about limited bandwidth. You've already done such a nice job of giving feedback. I occasionally have to explain this to reviewers: *Any* feedback is helpful to help improve a submission, and we work with your schedule to complete your review. COVID-19 has put all of us under major strain, so what's the hurry? Given that you offered comments, I do sense some interest. It would be great to have you on board!

I'm interpreting your response as "not available" but I am happy with slow progress vs. no progress. 👍🏾  Please follow up to let me know. I'll respect your final answer.
