+++ kalmus/utils/Artist.py
@@ -14,18 +14,19 @@ def compute_dominant_color(image, n_clusters=3, max_iter=10, threshold_error=1.0
     """
     Compute the dominant color of an input image using the kmeans clustering. The centers of the
     clusters are the dominant colors of the input image
-    :param image: input image. Either a multi-channel color image or a grayscale image (2D image)
+
+    :param image: input image. Either a multi-channel color image or a grayscale image (2D image) \
     Expected shape of image is height x width x (channels)
     :param n_clusters: number of clusters
     :param max_iter: maximum iterations before terminating kmeans clustering
-    :param threshold_error: threshold error for terminating kmeans clustering. kmean clustering terminate
-    when the errors between the current computed and previous computed cluster center is under this
+    :param threshold_error: threshold error for terminating kmeans clustering. kmean clustering terminate \
+    when the errors between the current computed and previous computed cluster center is under this \
     threshold
-    :param attempts: Number of attempts to rerun the Kmeans clustering with different initial cluster
+    :param attempts: Number of attempts to rerun the Kmeans clustering with different initial cluster \
     centers. Since kmeans clustering randomly choose n number of cluster in its initialization process.
-    :return: An array of n cluster centers and an array of relative size (in percentage) of the clusters.
-    Expected output shape: n_clusters x channels (cluster centers), n_clusters (relative size in percentage)
-    E.g. for an input color image with 3 channels and n_clusters = 3  Output can be [[255, 255, 255],
+    :return: An array of n cluster centers and an array of relative size (in percentage) of the clusters. \
+    Expected output shape: n_clusters x channels (cluster centers), n_clusters (relative size in percentage) \
+    E.g. for an input color image with 3 channels and n_clusters = 3  Output can be [[255, 255, 255], \
     [126, 75, 198], [186, 145, 122]], [0.4, 0.5, 0.1]
     """
     assert len(image.shape) >= 2, "The input must be a 2 dimensional image"
@@ -46,6 +47,7 @@ def flatten_image(image):
     """
     Flat the input 2D image into an 1D image while preserve the channels of the input image
     with shape==[height x width, channels]
+
     :param image: Input 2D image (either multi-channel color image or greyscale image)
     :return: The flatten 1D image. shape==(height x width, channels)
     """
@@ -60,10 +62,11 @@ def flatten_image(image):
 def compute_percents_of_labels(label):
     """
     Compute the ratio/percentage size of the labels in an labeled image
+
     :param label: the labeled 2D image
-    :return: An array of relative size of the labels in the image. Indices of the sizes in the array
-    is corresponding to the labels in the labeled image. E.g. output [0.2, 0.5, 0.3] means label 0's size
-    is 0.2 of the labeled image, label 1' size is 0.5 of the labeled image, and label 2's size is 0.3 of
+    :return: An array of relative size of the labels in the image. Indices of the sizes in the array \
+    is corresponding to the labels in the labeled image. E.g. output [0.2, 0.5, 0.3] means label 0's size \
+    is 0.2 of the labeled image, label 1' size is 0.5 of the labeled image, and label 2's size is 0.3 of \
     the labeled image.
     """
     # Get the bins of the histogram. Since the last bin of the histogram is [label, label+1]
@@ -81,12 +84,13 @@ def compute_percents_of_labels(label):
 def compute_mode_color(image, bin_size = 10):
     """
     compute the mode color of an input image
-    :param image: either a multi-channel color image or a single channel greyscale image
+
+    :param image: either a multi-channel color image or a single channel greyscale image \
     Expected shape of image: height x width x (channels)
-    :param bin_size: Histogramize the input image. Color/intensity of each pixel in the
-    image will be accumulate in the bins with size==bin_size. The output mode color/intensity
[…]
