+++ api_docs_source/utils/measure_utils.rst
@@ -1,7 +1,0 @@
-Measure Utility
-*****************
-
-.. automodule:: kalmus.utils.measure_utils
-    :members: nrmse_similarity, ssim_similarity, cross_correlation, local_cross_correlation, compare_needleman_wunsch, compare_smith_waterman, generate_hue_strings_from_color_barcode, generate_brightness_string_from_brightness_barcode
-    :undoc-members:
-    :show-inheritance:
