+++ api_docs_source/index.rst
@@ -1,21 +1,0 @@
-.. KALMUS documentation master file, created by
-   sphinx-quickstart on Wed Feb 24 16:47:39 2021.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
-Welcome to KALMUS's documentation!
-==================================
-
-.. toctree::
-   :maxdepth: 2
-   :caption: Contents:
-
-   Home <self>
-   API Documentation <kalmus>
-
-Indices and tables
-==================
-
-* :ref:`genindex`
-* :ref:`modindex`
-* :ref:`search`
