+++ api_docs_source/barcodes/BarcodeGenerator.rst
@@ -1,7 +1,0 @@
-BarcodeGenerator Class
-*************************
-
-.. automodule:: kalmus.barcodes.BarcodeGenerator
-    :members: BarcodeGenerator, build_barcode_from_json
-    :undoc-members:
-    :show-inheritance:
