+++ kalmus/tkinter_windows/MainWindowVersion2.py
@@ -11,7 +11,7 @@ from kalmus.tkinter_windows.SaveBarcodeWindow import SaveBarcodeWindow
 from kalmus.tkinter_windows.LoadStackWindow import LoadStackWindow
 from kalmus.tkinter_windows.LoadJsonWindow import LoadJsonWindow
 from kalmus.tkinter_windows.ReshapeBarcodeWindow import ReshapeBarcodeWindow
-from kalmus.tkinter_windows.KALMUS_utils import paint_hue_hist, update_axes_title, resource_path
+from kalmus.tkinter_windows.gui_utils import paint_hue_hist, update_axes_title, resource_path
 from kalmus.tkinter_windows.plot_barcodes_windows.WhichBarcodeInpsectWindow import WhichBarcodeInspectWindow
 from kalmus.tkinter_windows.StatsInfoWindow import StatsInfoWindow
 from kalmus.tkinter_windows.SaveImageWindow import SaveImageWindow
@@ -46,6 +46,7 @@ class MainWindow():
     def __init__(self, barcode_tmp, barcode_gn, figsize=(12, 5), dpi=100):
         """
         Initialize
+
         :param barcode_tmp: The temporary barcode object for software initialization
         :param barcode_gn: The barcode generator object
         :param figsize: The size of the plotted figure
@@ -194,6 +195,7 @@ class MainWindow():
         close the Mainwindow
         Check if the Generate Barcode window is still open before quiting the Main program
         Return (cancel the quit) if the Generate Barcode window is still open.
+
         :return:
         """
         # Check if generate barcode window is opened
@@ -213,6 +215,7 @@ class MainWindow():
     def quit(self):
         """
         Quit the main window
+
         :return:
         """
         self.root.quit()
@@ -221,6 +224,7 @@ class MainWindow():
     def check_meta_info(self):
         """
         Instantiate the WhichBarcodeCheckMeta window
+
         :return:
         """
         WhichBarcodeCheckMeta(self.barcode_1, self.barcode_2, self.barcodes_stack)
@@ -228,6 +232,7 @@ class MainWindow():
     def show_barcode(self):
         """
         Instantiate the WhichBarcodeInspectWindow
+
         :return:
         """
         WhichBarcodeInspectWindow(self.barcode_1, self.barcode_2, dpi=100, figsize=(7, 4))
@@ -235,6 +240,7 @@ class MainWindow():
     def load_json_barcode(self):
         """
         Instantiate the LoadJsonWindow
+
         :return:
         """
         LoadJsonWindow(self.barcode_gn, self.barcode_1, self.barcode_2, self.ax,
@@ -243,6 +249,7 @@ class MainWindow():
     def load_stack_barcode(self):
         """
         Instantiate the LoadStackWindow
+
         :return:
         """
         LoadStackWindow(self.barcodes_stack, self.barcode_1, self.barcode_2, self.ax,
@@ -251,6 +258,7 @@ class MainWindow():
     def reshape_barcode(self):
         """
         Instantiate the ReshapeBarcodeWindow
+
         :return:
         """
         ReshapeBarcodeWindow(self.barcode_1, self.barcode_2, self.ax, self.canvas)
@@ -258,6 +266,7 @@ class MainWindow():
     def stats_info(self):
         """
         Instantiate the StatsInfoWindow
+
         :return:
         """
         StatsInfoWindow(self.barcode_1, self.barcode_2)
@@ -265,6 +274,7 @@ class MainWindow():
     def generate_barcode(self):
         """
         Instantiate the GenerateBarcodeWindow
+
         :return:
         """
         if not self.generate_window_opened:
@@ -277,6 +287,7 @@ class MainWindow():
     def save_barcode_on_stack(self):
         """
         Instantiate the SaveBarcodeWindow
+
[…]
