+++ kalmus/tkinter_windows/GenerateBarcodeWindow.py
@@ -9,7 +9,7 @@ import os
 import threading
 
 from kalmus.tkinter_windows.meta_info_windows.SpecifyMetaDataWindow import SpecifyMetaDataWindow
-from kalmus.tkinter_windows.KALMUS_utils import resource_path
+from kalmus.tkinter_windows.gui_utils import resource_path
 
 
 class GenerateBarcodeWindow():
@@ -21,6 +21,7 @@ class GenerateBarcodeWindow():
     def __init__(self, barcode_generator, barcode_stack):
         """
         Initialize
+
         :param barcode_generator: The barcode generator
         :param barcode_stack: The dictionary that stores all the barcode on the memory
         """
@@ -250,6 +251,7 @@ class GenerateBarcodeWindow():
     def quit(self):
         """
         Quit the main window
+
         :return:
         """
         self.window.quit()
@@ -258,6 +260,7 @@ class GenerateBarcodeWindow():
     def specify_data(self):
         """
         Instantiate the SpecifyMetaDataWindow
+
         :return:
         """
         SpecifyMetaDataWindow(self.meta_data_dict)
@@ -265,6 +268,7 @@ class GenerateBarcodeWindow():
     def update_thread_entry(self):
         """
         Enable or disable and change the label prompt when user check/uncheck the multi-thread checkbox
+
         :return:
         """
         if self.var_multi_thread.get() == 0:
@@ -277,6 +281,7 @@ class GenerateBarcodeWindow():
     def update_rescale_entry(self):
         """
         Enable or disable and change the label prompt when user check/uncheck the rescale frames checkbox
+
         :return:
         """
         if self.var_rescale_frame.get() == 0:
@@ -289,6 +294,7 @@ class GenerateBarcodeWindow():
     def update_save_frame_entry(self):
         """
         Enable or disable the text entry when user check/uncheck the save frames checkbox
+
         :return:
         """
         if self.var_saved_frame.get() == 0:
@@ -299,6 +305,7 @@ class GenerateBarcodeWindow():
     def time_unit(self):
         """
         Change the acquisition unit to Time when user switch to the time radio button
+
         :return:
         """
         self.skip_over_label['text'] = "Start at (mins:secs): "
@@ -308,6 +315,7 @@ class GenerateBarcodeWindow():
     def frame_unit(self):
         """
         Change the acquisition unit to Frame when user switch to the frame radio button
+
         :return:
         """
         self.skip_over_label['text'] = "Start at (frames): "
@@ -317,6 +325,7 @@ class GenerateBarcodeWindow():
     def browse_folder(self):
         """
         Browse the folder
+
         :return:
         """
         # Get the file name from the user specification
@@ -332,6 +341,7 @@ class GenerateBarcodeWindow():
         """
         Generate the barcode in a another thread
         to avoid the frozen tkinter window issue
+
         :return:
         """
         threading.Thread(target=self.generate_barcode).start()
@@ -339,6 +349,7 @@ class GenerateBarcodeWindow():
     def acquire_generation_param(self):
         """
         Acquire the barcode generation parameters
+
         :return:
         """
         # Get barcode type, frame sampling type, and color/brightness metric
@@ -427,6 +438,7 @@ class GenerateBarcodeWindow():
     def generate_barcode(self):
         """
         Generate the barcode using the given parameters
+
         :return:
         """
         self.disable_generate_button()
@@ -575,6 +587,7 @@ class GenerateBarcodeWindow():
     def disable_generate_button(self):
         """
         Disable the generate button and Specify Meta Data button once generation starts
+
         :return:
         """
         # Update the generate button text to the processing
@@ -585,6 +598,7 @@ class GenerateBarcodeWindow():
     def enable_generate_button(self):
         """
[…]
