+++ kalmus/tkinter_windows/SaveBarcodeWindow.py
@@ -5,7 +5,7 @@ import tkinter.filedialog
 from tkinter.messagebox import showwarning, showinfo
 import os
 
-from kalmus.tkinter_windows.KALMUS_utils import resource_path
+from kalmus.tkinter_windows.gui_utils import resource_path
 
 
 class SaveBarcodeWindow():
