+++ kalmus/utils/artist.py
@@ -81,7 +81,7 @@ def compute_percents_of_labels(label):
     return percent_of_dominance
 
 
-def compute_mode_color(image, bin_size = 10):
+def compute_mode_color(image, bin_size=10):
     """
     compute the mode color of an input image
 
@@ -139,8 +139,8 @@ def compute_brightest_color_and_brightness(grey_image, color_image, return_min=F
         grey_image = cv2.GaussianBlur(grey_image.copy(), (blur_radius, blur_radius), 0)
     min_brightness, max_brightness, min_loc, max_loc = cv2.minMaxLoc(grey_image)
     if return_min:
-        return color_image[max_loc[::-1]], max_brightness, max_loc[::-1],\
-                color_image[min_loc[::-1]], min_brightness, min_loc[::-1]
+        return color_image[max_loc[::-1]], max_brightness, max_loc[::-1], \
+               color_image[min_loc[::-1]], min_brightness, min_loc[::-1]
     else:
         return color_image[max_loc[::-1]], max_brightness, max_loc[::-1]
 
@@ -178,7 +178,7 @@ def find_bright_spots(image, n_clusters=3, blur_radius=21, amount_of_bright_part
     threshed_img = cv2.dilate(threshed_img, None, iterations=4)
 
     # Get the location of all white pixel in binary threshold image
-    locs = np.argwhere(threshed_img==255)
+    locs = np.argwhere(threshed_img == 255)
     try:
         # convert to np.float32
         Z = np.float32(locs)
@@ -259,7 +259,8 @@ def random_sample_pixels(img, sample_ratio=0, mode="row-col"):
     return np.array(random_pixels)
 
 
-def watershed_segmentation(image, minimum_segment_size=0.0004, base_ratio=0.5, denoise_disk_size = 5, gradiant_disk_size = 5,
+def watershed_segmentation(image, minimum_segment_size=0.0004, base_ratio=0.5, denoise_disk_size=5,
+                           gradiant_disk_size=5,
                            marker_disk_size=15):
     """
     Label the connected regions in an image.
@@ -486,7 +487,7 @@ def _sRGB2luminance(sRGB):
     lRGB = np.zeros(shape=sRGB.shape)
     lRGB[sRGB <= 0.03928] = sRGB[sRGB <= 0.03928] / 12.92
     lRGB[sRGB > 0.03928] = ((sRGB[sRGB > 0.03928] + 0.055) / 1.055) ** 2.4
-    luminance = 0.2126 * lRGB[...,0] + 0.7152 * lRGB[...,1] + 0.0722 * lRGB[...,2]
+    luminance = 0.2126 * lRGB[..., 0] + 0.7152 * lRGB[..., 1] + 0.0722 * lRGB[..., 2]
     return luminance
 
 
@@ -547,6 +548,130 @@ def grabcut_foreback_segmentation(image, start_row=0, start_col=0, row_size=-1, 
     background_image = image[np.where((mask == 0) | (mask == 2))]
 
     return foreground_image, background_image
+
+
+def find_letter_box_from_videos(video, num_sample=30):
+    """
+    Find the position of letterbox (black bars around the scene) in the input cv2 video object.
+    The function samples out num_sample (20 by default) number of frames from cv2 video object, and find the position
+    of letterbox in each frame. The function uses the median results from all sampled frames as the final position of
+    video's letterbox.
+
+    Notice that the function assumes the letterbox is black or very close to black.
+
+    :param video: Input video object captured by cv2.VideoCapture()
+    :param num_sample: Number of frames to sample from video for finding letterbox
+    :return: The smaller row index of letterbox (bound for upper horizontal letterbox), \
+             the larger row index of letterbox (bound for lower horizontal letterbox), \
[…]
