### 2021-03-02 88f97f7
fix the 404 error on kalmus.utils.artist documentation page
