+++ automated_tests/conftest.py
@@ -1,0 +1,5 @@
+import pytest
+
+
+
+#@pytest.fixture(scope="session")
