+++ automated_tests/test_artist.py
@@ -3,21 +3,26 @@ import pytest
 import numpy as np
 import matplotlib.pyplot as plt
 import cv2
+import warnings
+
+# Filter out the warnings raised by numpy.dtype and numpy.ufunc
+# See pull requests on NumPy https://github.com/numpy/numpy/pull/432 for references about these warnings
+# See https://stackoverflow.com/questions/40845304/runtimewarning-numpy-dtype-size-changed-may-indicate-binary-incompatibility
+# for how the warnings are dealt in this automated test suite
+warnings.filterwarnings("ignore", message="numpy.dtype size changed")
+warnings.filterwarnings("ignore", message="numpy.ufunc size changed")
 
 # kalmus module being tested
 import kalmus.utils.artist as artist
 
 
-@pytest.fixture(scope="session")
-def get_test_color_image():
-    return plt.imread("automated_tests/test_data/test_color_image.jpg", format='jpeg')
+def get_test_color_video():
+    video = cv2.VideoCapture("automated_tests/test_data/test_color_video.mp4")
+    return video
 
 
-@pytest.fixture(scope="session")
-def get_test_gray_image():
-    clr_image = plt.imread("automated_tests/test_data/test_color_image.jpg", format='jpeg')
-    gray_image = cv2.cvtColor(clr_image, cv2.COLOR_RGB2GRAY)
-    return gray_image
+def get_test_color_image_with_letterbox():
+    return plt.imread("automated_tests/test_data/test_color_image_letterboxing.jpg", format="jpeg")
 
 
 def test_compute_mean_color(get_test_color_image):
@@ -102,3 +107,91 @@ def test_random_sample_pixels(get_test_color_image):
     # Potential small loses in size when resampling discrete/integer number of pixels
     epsilon = 1e-5
     assert image.size * (sample_ratio_whole - epsilon) < sampled_1D.size < image.size * (sample_ratio_whole + epsilon)
+
+
+def test_watershed_segmentation(get_test_color_image):
+    color_image = get_test_color_image
+    segmented_image, gray_image = artist.watershed_segmentation(color_image)
+    assert segmented_image.shape == gray_image.shape == color_image.shape[:2]
+
+
+def test_color_of_regions(get_test_color_image):
+    color_image = get_test_color_image
+    segmented_image, gray_image = artist.watershed_segmentation(color_image)
+    avg_colors, brightest_colors, region_size_ratio = artist.color_of_regions(segmented_image, color_image)
+
+    assert np.array(avg_colors).shape[0] == np.unique(segmented_image).size
+    assert np.array(brightest_colors).shape[0] == np.unique(segmented_image).size
+    # Epsilon for small precision that may be lost in the floating point computation
+    epsilon = 1e-6
+    assert 1.0 - epsilon < np.sum(region_size_ratio) < 1.0 + epsilon
+
+
+@pytest.mark.filterwarnings("ignore::DeprecationWarning")
+def test_contrast_between_regions(get_test_color_image):
+    color_image = get_test_color_image
+    segmented_image, gray_image = artist.watershed_segmentation(color_image)
+    avg_colors, brightest_colors, region_size_ratio = artist.color_of_regions(segmented_image, color_image)
+
+    rag = artist.get_rag(gray_image, segmented_image)
+    num_regions= np.unique(segmented_image).size
+    matrix = artist.rag_to_matrix(rag, num_regions=num_regions)
+
+    contrast_matrix = artist.contrast_between_regions(avg_colors, matrix)
+    assert contrast_matrix.shape == (num_regions, num_regions)
+    contrast_matrix_weighted = artist.contrast_between_regions(brightest_colors, matrix,
[…]
