### 2021-03-04 b7817af
update the automated test suite
