+++ kalmus/utils/visualization_utils.py
@@ -5,26 +5,32 @@ import matplotlib.pyplot as plt
 from kalmus.utils.artist import get_contrast_matrix_and_labeled_image
 
 
-def show_color(color, figure_size=(9, 6), title="Undefined Test",
-               axis_off=False, save_image=False, file_name="test.png"):
+def show_color(color, figure_size=(9, 6), title="",
+               axis_off=False, save_image=False, file_name="test.png", return_color=False):
     """
     Plot the rgb color.
 
     :param color: 1D array contains the R, G, and B channel values
+    :param figure_size: Size of the plotted figure
     :param title: The title of the plotted color
     :param axis_off: True to set the axis of the plot figure off. False to have the axis on the side of \
     figure.
     :param save_image: True to save the plot figure into the path that user provide. False not to save \
     the plot figure
     :param file_name: The filename of the saved figure. The default path to the saved image is test.png
+    :param return_color: Return the numpy array (shape==30 x 30 x 3) of color block if True. Directly plot the color if False
+    :return: Return the numpy array of color block if return_color is True.
     """
     # Generate a block image where each pixel of this block image is the plotted RGB color
     clr = np.ones(shape=(30, 30, 3), dtype=np.uint8) * np.uint8(color)
+
+    if return_color:
+        return clr
+
     # Set up the figure
     plt.figure(figsize=figure_size)
     plt.imshow(clr)
     plt.title(title)
-
     # Whether to set the axis off
     if axis_off:
         plt.axis('off')
@@ -34,8 +40,9 @@ def show_color(color, figure_size=(9, 6), title="Undefined Test",
     plt.show()
 
 
-def show_colors_in_sequence(colors, figure_size=(9, 6), title="Undefined Test", axis_off=False, \
-                            save_image=False, file_name="test.png", horizontal=True):
+def show_colors_in_sequence(colors, figure_size=(9, 6), title="", axis_off=False,
+                            save_image=False, file_name="test.png", horizontal=True,
+                            return_color_sequence=False):
     """
     Plot a sequence of RGB colors either horizontally or vertically in line
 
@@ -45,7 +52,10 @@ def show_colors_in_sequence(colors, figure_size=(9, 6), title="Undefined Test", 
     :param axis_off: True to set the axis of the figure off. False to have the axis with the figure
     :param save_image: True to save the plot figure. False do not save figure
     :param file_name: the path of the saved figure
-    :param horizontal: True to plot the sequence of colors horizontally, False to plot the sequence of colors vertically
+    :param horizontal: True to plot the sequence of colors horizontally, False to plot the sequence of colors \
+                       vertically
+    :param return_color_sequence: Return the numpy array (shape==30 x (30 * num_colors) x 3) of color sequence if True \
+                                  Directly plot the color sequence if False
     """
     # A chain/sequence of color blocks
     clr_chain = []
@@ -62,6 +72,9 @@ def show_colors_in_sequence(colors, figure_size=(9, 6), title="Undefined Test", 
     if horizontal:
         clr_chain = clr_chain.transpose([1, 0, 2])
 
+    if return_color_sequence:
+        return clr_chain
+
     plt.figure(figsize=figure_size)
     plt.imshow(clr_chain)
     plt.title(title)
@@ -73,7 +86,7 @@ def show_colors_in_sequence(colors, figure_size=(9, 6), title="Undefined Test", 
 
 
 def show_color_matrix(color_2d_array, mode="truncate", figure_size=(9, 4),
-                      title="Undefined Test", axis_off=False,
+                      title="", axis_off=False,
                       save_image=False, file_name="Block.png", return_matrix=False, return_figure=False):
     """
[…]
