+++ docs/_sources/utils/visualization_utils.rst.txt
@@ -2,6 +2,6 @@ Visualization Utility
 ***********************
 
 .. automodule:: kalmus.utils.visualization_utils
-    :members: show_color, show_colors_in_sequence, show_colors_in_cube, show_image, show_high_contrast_region, show_low_contrast_region, extract_region_with_index
+    :members: show_color, show_colors_in_sequence, show_colors_in_cube, show_image_aa, show_high_contrast_region, show_low_contrast_region, extract_region_with_index
     :undoc-members:
     :show-inheritance:

[…]
