### 2021-03-06 16738e2
update the full test suite with associated GitHub Action
