+++ tests/test_Barcode.py
@@ -1,6 +1,6 @@
 # Third-party modules
 import pytest
-import matplotlib.pyplot as plt
+import os
 import warnings
 import numpy as np
 import cv2
@@ -16,7 +16,17 @@ import kalmus.barcodes.Barcode as BarcodeClasses
 
 
 def get_template_barcode():
-    return BarcodeClasses.Barcode(color_metric="average", frame_type="whole_frame", barcode_type="Color")
+    return BarcodeClasses.Barcode(color_metric="Average", frame_type="Whole_frame")
+
+
+def get_template_color_barcode():
+    return BarcodeClasses.ColorBarcode(color_metric="Average", frame_type="Whole_frame",
+                                       barcode_type="Color")
+
+
+def get_template_brightness_barcode():
+    return BarcodeClasses.BrightnessBarcode(color_metric="Average", frame_type="Whole_frame",
+                                            barcode_type="Brightness")
 
 
 def test_foreback_segmentation(get_test_color_image):
@@ -27,8 +37,8 @@ def test_foreback_segmentation(get_test_color_image):
 
 
 def test_Barcode():
-    barcode = BarcodeClasses.Barcode(color_metric="average", frame_type="whole_frame", sampled_frame_rate=2,
-                                     skip_over=10, total_frames=100, barcode_type="color")
+    barcode = BarcodeClasses.Barcode(color_metric="Average", frame_type="Whole_frame", sampled_frame_rate=2,
+                                     skip_over=10, total_frames=100, barcode_type="Color")
     assert barcode.color_metric == "Average"
     assert barcode.frame_type == "Whole_frame"
     assert barcode.barcode_type == "Color"
@@ -46,7 +56,7 @@ def test_read_video():
 
     barcode.total_frames = 1e6
     barcode.save_frames_in_generation = True
-    barcode.enable_save_frames()
+    barcode.enable_save_frames(sampled_rate=-1)
     barcode.read_video("tests/test_data/test_color_video.mp4")
     assert barcode.total_frames == barcode.film_length_in_frames - barcode.skip_over
     assert barcode.high_bound_hor != 0
@@ -58,7 +68,9 @@ def test_find_film_letter_box():
     barcode = get_template_barcode()
     barcode.read_video("tests/test_data/test_color_video.mp4")
     barcode.find_film_letterbox()
+    assert barcode.low_bound_hor >= 0
     assert barcode.high_bound_hor != 0
+    assert barcode.low_bound_ver >= 0
     assert barcode.high_bound_ver != 0
 
 
@@ -109,6 +121,7 @@ def test_process_frame(get_test_color_image):
 
     barcode.frame_type = "Background"
     processed_frame = barcode.process_frame(color_frame)
+    barcode.enable_rescale_frames_in_generation(rescale_factor=0.6)
     assert processed_frame.shape[-1] == color_frame.shape[-1]
     assert len(processed_frame.shape) == 2
 
@@ -125,3 +138,198 @@ def test_resize_frame(get_test_color_image):
            color_image.shape[0] * (barcode.rescale_frame_factor + epsilon)
     assert color_image.shape[1] * (barcode.rescale_frame_factor - epsilon) < resize_frame.shape[1] < \
            color_image.shape[1] * (barcode.rescale_frame_factor + epsilon)
+
+
+def test_get_color_from_frame(get_test_color_image):
+    color_frame = get_test_color_image
+    barcode = get_template_barcode()
+
+    barcode.color_metric = "Average"
+    color = barcode.get_color_from_frame(color_frame)
+    assert color.shape == (3,)
+
+    barcode.color_metric = "Median"
+    color = barcode.get_color_from_frame(color_frame)
+    assert color.shape == (3,)
+
[…]
