+++ .github/workflows/test-codecov.yml
@@ -30,6 +30,7 @@ jobs:
       run: |
         pip install pytest
         pip install pytest-cov
+        pip install .
         python -m pytest tests --cov=kalmus --cov-config=.coveragerc --cov-report=xml
     - name: Upload coverage to Codecov
       uses: codecov/codecov-action@v1
