+++ .github/workflows/test-codecov.yml
@@ -16,7 +16,7 @@ jobs:
     strategy:
       matrix:
         python-version: ["3.7", "3.8"]
-        os: [ubuntu-latest, macos-latest, windows-latest]
+        os: [windows-latest]
     env:
       OS: ${{ matrix.os }}
       PYTHON: ${{ matrix.python-version }}
@@ -35,11 +35,11 @@ jobs:
     - name: Upload coverage to Codecov
       uses: codecov/codecov-action@v1
       with:
-        files: ./coverage1.xml,./coverage2.xml
+        files: ./coverage.xml
         directory: ./coverage/reports/
         flags: unittests
         env_vars: OS,PYTHON
         name: codecov-umbrella
         fail_ci_if_error: true
-        path_to_write_report: ./coverage/codecov_report.txt
+        path_to_write_report: ./codecov_report.txt
         verbose: true
