### 2021-03-06 48c60fc
clean out the saved code coverage results
