### 2021-03-06 9c6175d
extend the running os of test and build actions
