### 2021-03-06 3c49778
improve the auto tests on kalmus.utils.visualization_utils
