+++ kalmus/__init__.py
@@ -1,1 +1,1 @@
-__version__ = "kalmus 1.3.7"
+__version__ = "1.3.7"
