### 2021-03-07 fb98a05
update the release 1.3.7 with command-line barcode generation pipeline
