+++ docs/_sources/update_log.rst.txt
@@ -4,6 +4,23 @@ KALMUS Update Log (v1.3.6)
 Update Log
 -----------------
 
+update 1.3.7
+^^^^^^^^^^^^^^^^
+
+**Changes**
+
+* The kalmus.utils.Artist module is renamed to kalmus.utils.artist
+
+**Add features**
+
+* New console script `kalmus-generator` that allows users to automate their barcode generation process through this command line interface. Example usage::
+
+    $ kalmus-generator -p tests/test_data/test_color_video.mp4 --frame_type Whole_frame --color_metric Average --skip 10 --step 1 --total_frames 100 --barcode_type Color
+
+* Use ``$ kalmus-generator -h`` to check the description of available arguments in details.
+
+---------------------------
+
 Update 1.3.6
 ^^^^^^^^^^^^^^^^
 
