+++ README.md
@@ -2,7 +2,7 @@
 [![Python Version](https://img.shields.io/pypi/pyversions/kalmus.svg)](https://pypi.org/project/kalmus/)
 [![PyPI Version](https://img.shields.io/pypi/v/kalmus.svg)](https://pypi.org/project/kalmus/)
 [![codecov](https://codecov.io/gh/KALMUS-Color-Toolkit/KALMUS/branch/master/graph/badge.svg)](https://codecov.io/gh/KALMUS-Color-Toolkit/KALMUS)
-[![](https://img.shields.io/pypi/l/kalmus.svg)](https://pypi.org/project/kalmus/)
+[![License](https://img.shields.io/pypi/l/kalmus.svg)](https://pypi.org/project/kalmus/)
 ![codecov workflow](https://github.com/KALMUS-Color-Toolkit/KALMUS/actions/workflows/test-codecov.yml/badge.svg)
 ![build workflow](https://github.com/KALMUS-Color-Toolkit/KALMUS/actions/workflows/python-package.yml/badge.svg)
 
@@ -14,6 +14,13 @@ This package serves two purposes: (1) various ways to measure, calculate and com
 and (2) various ways to visualize a film's color. We have named the software KALMUS in homage to 
 Natalie Kalmus (1882 - 1965), a Technicolor Director who oversaw the color palettes of nearly 300 
 Hollywood feature films.
+
+- What is a Color Barcode: [KALMUS: tools for color analysis of films](paper/joss-paper.md)   
+- How can I install the KALMUS: [KALMUS Installation Guide](https://kalmus-color-toolkit.github.io/KALMUS/install.html) 
+and [KALMUS PyPI Homepage](https://pypi.org/project/kalmus/).
+- How do I contribute to the KALMUS: [KALMUS Contribution Guidelines](CODE_OF_CONDUCT.md)
+- How to run KALMUS's automated test suite: [Auomated Test Suite](tests/)
+
 
 # API Documentation
 
@@ -51,157 +58,28 @@ Once the package is installed. The functionalities of the KALMUS are accessible 
 >>> 1.3.7 
 ```
 
+# Contribution
+
+We encourage contributions, including bug fixes and new features, from our community users. When contributing to the 
+kalmus package, please first contact the project maintainers by [email](yc015@bucknell.edu) or opening an issue. If 
+your bug fixex or new features changing the current behavior of package, please specify the changed behaviors and 
+reasons for changing in the discussion with project maintainers. 
+
+We encourage inclusive and friendly discussion. Please follow our [code of conduct](CODE_OF_CONDUCT.md) when 
+communicating. 
+
 # Precomputed Barcodes
 
-Precomputed barcodes are accessible upon requests. Please email the Project maintainer Yida Chen <yc015@bucknell.edu> about your needs. 
+Precomputed barcodes are accessible upon requests. Please email the Project maintainer Yida Chen <yc015@bucknell.edu> 
+about your needs. 
 
 # Acknowledgment
 
-The authors wish to thank the Mellon Foundation, the Dalal Family Foundation, and the Bucknell University Humanities Center for their support on this project.
+The authors wish to thank the Mellon Foundation, the Dalal Family Foundation, and the Bucknell University Humanities 
+Center for their support on this project.
 The project is released under the open-source MIT License.
 
 # Update Log
 
-### update 1.3.7
-**Changes**
-- The kalmus.utils.Artist module is renamed to kalmus.utils.artist
-
-**Add features**
-- New console script `kalmus-generator` that allows users to automate their barcode generation process
-through this command line interface. Example usage:
-```
[…]
