+++ README.md
@@ -61,7 +61,7 @@ Once the package is installed. The functionalities of the KALMUS are accessible 
 # Contribution
 
 We encourage contributions, including bug fixes and new features, from our community users. When contributing to the 
-kalmus package, please first contact the project maintainers by [email]<yc015@bucknell.edu> or opening an issue. If 
+kalmus package, please first contact the project maintainers by email <yc015@bucknell.edu> or opening an issue. If 
 your bug fixex or new features changing the current behavior of package, please specify the changed behaviors and 
 reasons for changing in the discussion with project maintainers. 
 
