### 2021-03-07 #2980 @thomakra
Hei George,

Sorry for the late reply. I also have very limited time these days, but
since you say you are happy with slow progress I can try to have a look
over the coming few weeks.

-Thomas-

On Sun, 21 Feb 2021 at 22:43, George K. Thiruvathukal <
notifications@github.com> wrote:

> Thanks, @stsievert <https://github.com/stsievert>! Appreciate your
> initial input.
>
> @yc015 <https://github.com/yc015> I still don't have any reviewers, but I
> see the documentation issue reported by @stsievert
> <https://github.com/stsievert> as a blocker here. Can you please work to
> address this right away? And please suggest other reviewers from our
> reviewer list.
>
