+++ notebooks/README.md
@@ -1,0 +1,17 @@
+[![Project Status](https://img.shields.io/pypi/status/kalmus.svg)](https://pypi.org/project/kalmus/)
+[![Python Version](https://img.shields.io/pypi/pyversions/kalmus.svg)](https://pypi.org/project/kalmus/)
+[![PyPI Version](https://img.shields.io/pypi/v/kalmus.svg)](https://pypi.org/project/kalmus/)
+
+# Notebooks Tutorial
+This folder contains the Ipython notebooks that walk you through how to interact with KALMUS package.
+
+To run these notebooks, make sure you have install thhe latest version of Jupyter Lab 
+([Installation Instruction for JupyterLab](https://jupyterlab.readthedocs.io/en/stable/getting_started/installation.html)) 
+and the latest version of KALMUS package ([Installation Guide for KALMUS](https://kalmus-color-toolkit.github.io/KALMUS/install.html)).
+ 
+When running the notebooks, make sure you have cloned the project on your local machine or download the whole 
+[notebooks folder](../notebooks). Then, start the Jupyter Lab inside the notebooks folder using command:
+
+```
+$ jupyter lab
+```
