+++ notebooks/README.md
@@ -5,7 +5,7 @@
 # Notebooks Tutorial
 This folder contains the Ipython notebooks that walk you through how to interact with KALMUS package.
 
-To run these notebooks, make sure you have install thhe latest version of Jupyter Lab 
+To run these notebooks, make sure you have install the latest version of Jupyter Lab 
 ([Installation Instruction for JupyterLab](https://jupyterlab.readthedocs.io/en/stable/getting_started/installation.html)) 
 and the latest version of KALMUS package ([Installation Guide for KALMUS](https://kalmus-color-toolkit.github.io/KALMUS/install.html)).
  
