### 2021-03-08 74fb145
update the api tutorial notebook
