### 2021-03-08 daf0f5f
fix the figure not displayed error in api and gui tutorial notebooks
