+++ docs/_sources/install.rst.txt
@@ -8,7 +8,6 @@ python==3.7, 3.8, you can install the kalmus using pip (recommended)::
 
     $ pip install kalmus
 
-
 Alternatively, you could install the kalmus locally by first cloning the GitHub repo of Kalmus (`GitHub page <https://github.com/KALMUS-Color-Toolkit/KALMUS>`_).
 Then, move to the top directory of cloned kalmus project folder and install using pip command::
 
@@ -18,3 +17,12 @@ In both methods, the package's dependencies will be automatically installed. You
 installed in your environment using pip command::
 
     $ pip show kalmus
+
+Alternatively, in the version 1.3.7 and above, you can check the version of intsalled kalmus using its
+``.__version__`` attribute
+
+.. code-block:: python
+
+    >>> import kalmus
+    >>> print(kalmus.__version__) # Warning: The __version__ attribute is not available in the kalmus v.1.3.6 and backward
+    >>> 1.3.7
