### 2021-03-09 f077613
update the Markdown tutorial for Command-line interface
