+++ notebooks/README.md
@@ -3,9 +3,11 @@
 [![PyPI Version](https://img.shields.io/pypi/v/kalmus.svg)](https://pypi.org/project/kalmus/)
 
 # Notebooks Tutorial
-This folder contains the Ipython notebooks that walk you through how to interact with KALMUS package.
+Welcome to the KALMUS's notebook tutorials!
 
-To run these notebooks, make sure you have install the latest version of Jupyter Lab 
+This folder contains the IPython notebooks that walk you through how to interact with the KALMUS package.
+
+To run these notebooks, please install the latest version of Jupyter Lab 
 ([Installation Instruction for JupyterLab](https://jupyterlab.readthedocs.io/en/stable/getting_started/installation.html)) 
 and the latest version of KALMUS package ([Installation Guide for KALMUS](https://kalmus-color-toolkit.github.io/KALMUS/install.html)).
  
@@ -15,3 +17,13 @@ When running the notebooks, make sure you have cloned the project on your local 
 ```
 $ jupyter lab
 ```
+
+**Content:**  
+- [Notebook Figures](notebook_figures)
+- [Notebook Example Data](notebook_example_data): You are welcome to put your own data inside this folder when walking 
+through the examples in notebooks!
+- [Notebook Utility](notebook_utils.py): Plot utility for you to visualize certain processes occurring during the 
+barcode generation.
+- Notebook Guide for how to interact with [Graphic User Interface](user_guide_for_kalmus_gui.ipynb)
+- Notebook Guide for how to interact with [Application Programming Interface](user_guide_for_kalmus_api.ipynb)
+- Notebook Guide for how to interact with [Command-line Interface](USAGE_COMMAND_LINE_UI.md)
