+++ notebooks/USAGE_COMMAND_LINE_UI.md
@@ -1,0 +1,162 @@
+# Welcome to the Interactive User Guide for KALMUS (Command-line Interface)!
+
+In this Markdown Tutorial I will introduce:
+- **Installation of KALMUS package**
+- **How to use the Command-line interface to automate your barcode generation workflow**
+- **What are the available argument options for command-line generator**
+    - -p --path
+    - --color_metric
+    - --frame_type
+    - --barcode_type
+    - --skip
+    - -s --step
+    - -t --total_frames
+    - --num_thread
+    - --save_frame_rate
+    - --rescale_frame_factor
+    - -o --output_path
+
+# Notice
+
+The Command-line Interface of KALMUS is used specifically for **generating barcode** and **saving** it into 
+reloadable (via both [API](user_guide_for_kalmus_api.ipynb) and [GUI](user_guide_for_kalmus_gui.ipynb)) JSON object.
+
+We recommend our users to first walk through the IPython notebook Tutorials on either 
+[KALMUS's Graphic user interface](user_guide_for_kalmus_gui.ipynb) or [KALMUS's Application programming interface](user_guide_for_kalmus_api.ipynb), 
+in which the full functionality of KALMUS is covered.
+
+Once you have been familiar with the API, GUI, or both interfaces, you could further automate your barcode generation workflow 
+through this **Command-line interface**.
+
+# Installation Guide
+
+There are two ways that you could install KALMUS on you local machine:  
+1. (**Recommended**) Get the latest distribution of KALMUS from PyPI ([KALMUS Project Page on PyPI](https://pypi.org/project/kalmus/)).  
+Use command `$ pip install kalmus` or `$ pip install --upgrade kalmus` (if kalmus has been installed) to install the latest version of KALMUS package. All dependencies should be automatically installed during this process.
+
+2. Alternatively, you could install the KALMUS locally by first cloning the GitHub repo of Kalmus ([GitHub page](https://github.com/KALMUS-Color-Toolkit/KALMUS)). Then, move to the top directory of cloned KALMUS project and install using command `pip install .`. 
+
+**See our [Installation Guide](https://kalmus-color-toolkit.github.io/KALMUS/install.html) for more details.**
+
+Once the package is installed, you could verify the version of KALMUS package using command `$ pip show kalmus`  
+![KALMUS version](notebook_figures/kalmus_version.png)
+
+Alternatively, in version 1.3.7 and above, you can check the version of installed kalmus using its 
+`.__version__` attribute.
+
+```jupyter
+>>> import kalmus
+>>> print(kalmus.__version__) # Warning: The __version__ attribute is not available in the kalmus v.1.3.6 and backward
+>>> 1.3.7 
+```
+
+## Important!
+
+The **Command-line interface** is a new feature added into the KALMUS in its 1.3.7 version. To use Command-line 
+interface, you have to make sure the version of installed KALMUS is 1.3.7 or onward.
+
+# How do I use the command-line interface?
+
+The command-line interface is similar to the **BarcodeGenerator** object, which we have covered in [API Guide](user_guide_for_kalmus_api.ipynb). 
+You could invoke the command-line generator using command `$ kalmus-generator`. 
+
+Notice that if you have installed 
+kalmus>=1.3.7 but failed to start the kalmus-generator through this command, you could find the corresponding executable 
+in your_python_directory/Scripts/ directory on Windows or .local/bin/ directory on Linux.
+
+Check the available arguments for `kalmus-generator` command using the flag `-h`:
+
+```
+Windows OS
+$ kalmus-generator -h
+$ usage: kalmus-generator [-h] -p PATH --color_metric COLOR_METRIC --frame_type
[…]
