+++ CONTRIBUTING.md
@@ -1,4 +1,4 @@
 # Contributing
-We encourage contributions, including bug fixes and new features, from our community users. When contributing to the kalmus package, please first contact the project maintainers by email <yc015@bucknell.edu> or opening an issue. If your bug fixex or new features changing the current behavior of package, please specify the changed behaviors and reasons for changing in the discussion with project maintainers. 
+We encourage contributions, including bug fixes and new features, from our community users. When contributing to the kalmus package, please first contact the project maintainers by email <yc015@bucknell.edu> or opening an issue. If your bug fixes or new features changing the current behavior of the package, please specify the changed behaviors and reasons for changing in the discussion with project maintainers. 
 
 We encourage inclusive and friendly discussion. Please follow our [code of conduct](CODE_OF_CONDUCT.md) when communicating. 

[…]
