### 2021-03-09 69e5112
update the Markdown tutorials for KALMUS Command-line interface
