+++ docs/_sources/index.rst.txt
@@ -19,11 +19,19 @@ KALMUS Color Toolkit
 ======================
 Filmmakers, cinematographers, production designers, and then colorists (the person who fine tunes or grades the color in post production) spend enormous sums and labor deciding a film’s color palette. Why? Because color has an almost subliminal impact on how we perceive a film. Interpreting a film’s use of color, however, remains an elusive and understudied area. Despite how color shades a film’s meaning, the way audiences read color remains highly subjective. Moreover, as a time-based medium, understanding the patterning of color over the length of an entire film proves difficult for the typical viewer to see. KALMUS allows researchers to look at general patterns and trends in the use of color. It is useful to sometimes be able to process a large number of films, to quickly visualize their color palette and to compare the use of color in two films.
 
-KALMUS addresses these needs by (1) providing an interface that takes in either a video file (list formats here) or as a processed JSON file (files available for download), (2) allowing for the computation of a frame's "average" color in a number of ways, (3) allowing for the computation of a film's "average" color in a number of ways, (4) allowing for the segmentation of a frame in a number of ways and the subsequent computation of the color of various parts of the image, (5) providing an implementation of various ways to compare colors of two films, and (6) providing implementations of a few ways to visualize the color of a film, in particular, as color bar codes, as the 3D-plots in RBG-space and as histograms.
+KALMUS is a Python package for the computational analysis of colors in films. It provides quantitative tools to study and compare the use of film color. This package serves two purposes: (1) various ways to measure, calculate and compare a film's colors and (2) various ways to visualize a film's color. We have named the software KALMUS in homage to Natalie Kalmus (1882 - 1965), a Technicolor Director who oversaw the color palettes of nearly 300 Hollywood feature films.
+
+KALMUS utilizes the movie barcode as a visualization of the film's color. It has a modularized pipeline for the generation of barcodes using different measures of color and region of interest in each film frame. KALMUS provides a low-level API, high-level command line, and Graphic user interface for audience from all backgrounds to take advantage of its functionality.
+
+* What is a Movie Barcode: `KALMUS: tools for color analysis of films <https://github.com/KALMUS-Color-Toolkit/KALMUS/blob/master/paper/joss-paper.md>`_
+* KALMUS's project repository: `GitHub Repository for KALMUS project <https://github.com/KALMUS-Color-Toolkit/KALMUS>`_
+* KALMUS's latest release on PyPI: `KALMUS's PyPI Homepage <https://pypi.org/project/kalmus/>`_
+* Get started on KALMUS: `Notebook Tutorials for KALMUS'S API, GUI, and CLI <https://github.com/KALMUS-Color-Toolkit/KALMUS/tree/master/notebooks>`_
+* How do I contribute to the KALMUS: `KALMUS Contribution Guidelines <https://github.com/KALMUS-Color-Toolkit/KALMUS/blob/master/CONTRIBUTING.md>`_
 
 
 Indices and tables
-==================
+===================
 
 * :ref:`genindex`
 * :ref:`modindex`
