+++ README.md
@@ -89,7 +89,7 @@ communicating.
 
 We provide an automated test suite that covers the core functionality of KALMUS. Before running the automated test suite locally, 
 make sure you have installed the latest versions of [pytest](https://pypi.org/project/pytest/), [pytest-cov](https://pypi.org/project/pytest-cov/), 
-and [kalmus](https://pypi.org/project/kalmus/), and you have cloned the project repository on the master branch. 
+and [kalmus](https://pypi.org/project/kalmus/), and you have cloned the project repository on master branch. 
 
 To run the test suite:  
 - Go to the top directory of cloned KALMUS project
