+++ README.md
@@ -23,7 +23,7 @@ KALMUS utilizes the movie barcode as a visualization of the film's color. It has
 - What is a Movie Barcode: [KALMUS: tools for color analysis of films](paper/joss-paper.md)   
 - How do I install the KALMUS: [KALMUS Installation Guide](https://kalmus-color-toolkit.github.io/KALMUS/install.html) 
 and [KALMUS PyPI Homepage](https://pypi.org/project/kalmus/).
-- How do I use the KALMUS: [Notebook Tutorials for KALMUS'S API, GUI, and CLI](notebooks)
+- How do I use the KALMUS: [Notebook Tutorials for KALMUS's API, GUI, and CLI](notebooks)
 - How do I contribute to the KALMUS: [KALMUS Contribution Guidelines](CODE_OF_CONDUCT.md)
 - How do I run the KALMUS's automated test suite: [Auomated Test Suite](tests/)
 
