### 2021-03-09 94edfc7
include automated test for command_line generator (CLI)
