### 2021-03-09 e321158
correct the title in notebooks/README.md and Markdown tutorial for CLI
