+++ notebooks/README.md
@@ -26,4 +26,4 @@ through the examples in notebooks!
 barcode generation.
 - Notebook Guide for how to interact with [Graphic User Interface](user_guide_for_kalmus_gui.ipynb)
 - Notebook Guide for how to interact with [Application Programming Interface](user_guide_for_kalmus_api.ipynb)
-- Notebook Guide for how to interact with [Command-line Interface](USAGE_COMMAND_LINE_UI.md)
+- Markdown Guide for how to interact with [Command-line Interface](USAGE_COMMAND_LINE_UI.md)
