### 2021-03-10 4a06e58
correct one error in the measure metrics table in Tutorials
