+++ PyPI_README.md
@@ -1,0 +1,36 @@
+[![Project Status](https://img.shields.io/pypi/status/kalmus.svg)](https://pypi.org/project/kalmus/)
+[![Python Version](https://img.shields.io/pypi/pyversions/kalmus.svg)](https://pypi.org/project/kalmus/)
+[![PyPI Version](https://img.shields.io/pypi/v/kalmus.svg)](https://pypi.org/project/kalmus/)
+[![codecov](https://codecov.io/gh/KALMUS-Color-Toolkit/KALMUS/branch/master/graph/badge.svg)](https://codecov.io/gh/KALMUS-Color-Toolkit/KALMUS)
+[![License](https://img.shields.io/pypi/l/kalmus.svg)](https://pypi.org/project/kalmus/)
+![codecov workflow](https://github.com/KALMUS-Color-Toolkit/KALMUS/actions/workflows/test-codecov.yml/badge.svg)
+![build workflow](https://github.com/KALMUS-Color-Toolkit/KALMUS/actions/workflows/python-package.yml/badge.svg)
+
+# KALMUS
+
+KALMUS is a Python package for the computational analysis of colors in films. 
+It provides quantitative tools to study and compare the use of film color. 
+This package serves two purposes: (1) various ways to measure, calculate and compare a film's colors 
+and (2) various ways to visualize a film's color. We have named the software KALMUS in homage to 
+Natalie Kalmus (1882 - 1965), a Technicolor Director who oversaw the color palettes of nearly 300 
+Hollywood feature films.
+
+KALMUS utilizes the movie barcode as a visualization of the film's color. It has a modularized pipeline for the
+ generation of barcodes using different measures of color and region of interest in each film frame. KALMUS provides
+ a low-level API, high-level command line, and Graphic user interface for audience from all backgrounds to take
+ advantage of its functionality. 
+
+- What is a Movie Barcode: [KALMUS: tools for color analysis of films](https://github.com/KALMUS-Color-Toolkit/KALMUS/blob/master/paper/joss-paper.md)   
+- How do I install the KALMUS: [KALMUS Installation Guide](https://kalmus-color-toolkit.github.io/KALMUS/install.html)
+- How do I use the KALMUS: [Notebook Tutorials for KALMUS's API, GUI, and CLI](https://github.com/KALMUS-Color-Toolkit/KALMUS/tree/master/notebooks)
+- How do I contribute to the KALMUS: [KALMUS Contribution Guidelines](https://github.com/KALMUS-Color-Toolkit/KALMUS/blob/master/CONTRIBUTING.md)
+
+# API Documentation
+
+The KALMUS API reference is now available on 
+[https://kalmus-color-toolkit.github.io/KALMUS/kalmus.html](https://kalmus-color-toolkit.github.io/KALMUS/kalmus.html).
+
+# Update Log
+
+The full update log (from [v1.3.0](https://pypi.org/project/kalmus/1.3.0/) to [v1.3.7](https://pypi.org/project/kalmus/)) 
+is now available on [https://kalmus-color-toolkit.github.io/KALMUS/update_log.html](https://kalmus-color-toolkit.github.io/KALMUS/update_log.html)
