### 2021-03-10 890de76
update the tests for measure_utils.ssim_similarity
