### 2021-03-10 b10ac21
Improve the documentation on GUI
