+++ kalmus/tkinter_windows/GenerateBarcodeWindow.py
@@ -2,6 +2,8 @@
 
 import cv2
 import tkinter
+import tkinter.ttk as ttk
+from ttkthemes import ThemedTk
 import tkinter.filedialog
 from tkinter.messagebox import showerror, showwarning, showinfo
 import copy
@@ -29,7 +31,9 @@ class GenerateBarcodeWindow():
         self.barcode_stack = barcode_stack
 
         # Initialize the window
-        self.window = tkinter.Tk()
+        self.window = ThemedTk(theme="equilux")
+
+        self.window.configure(bg='#464646')
         self.window.wm_title("Barcode Generator")
         self.window.iconbitmap(resource_path("kalmus_icon.ico"))
 
@@ -40,15 +44,15 @@ class GenerateBarcodeWindow():
         self.video = None
 
         # Label prompt for the generated barcode's barcode type
-        barcode_type_label = tkinter.Label(self.window, text="Barcode Type: ")
+        barcode_type_label = ttk.Label(self.window, text="Barcode Type: ")
         barcode_type_label.grid(row=0, column=0)
 
         # Label prompt for the frame sampling type
-        frame_type_label = tkinter.Label(self.window, text="Frame Type: ")
+        frame_type_label = ttk.Label(self.window, text="Frame Type: ")
         frame_type_label.grid(row=1, column=0)
 
         # Label prompt for the color/brightness metric
-        color_metric_label = tkinter.Label(self.window, text="Color metric: ")
+        color_metric_label = ttk.Label(self.window, text="Color metric: ")
         color_metric_label.grid(row=2, column=0)
 
         # Variable that stores the user's choice of generated barcode type
@@ -56,7 +60,7 @@ class GenerateBarcodeWindow():
         self.barcode_type_var.set("Color")
 
         # Dropdown menu for the barcode type selection
-        dropdown_bar_type = tkinter.OptionMenu(self.window, self.barcode_type_var, "Color", "Brightness")
+        dropdown_bar_type = ttk.OptionMenu(self.window, self.barcode_type_var, "Color", "Brightness")
         dropdown_bar_type.grid(row=0, column=1)
 
         # Variable that stores the user's choice of frame sampling type
@@ -64,7 +68,7 @@ class GenerateBarcodeWindow():
         self.frame_type_var.set("Whole_frame")
 
         # Dropdown menu for the frame sampling type selection
-        dropdown_frame_type = tkinter.OptionMenu(self.window, self.frame_type_var, "Whole_frame", "Low_contrast_region",
+        dropdown_frame_type = ttk.OptionMenu(self.window, self.frame_type_var, "Whole_frame", "Low_contrast_region",
                                                  "High_contrast_region", "Foreground", "Background")
         dropdown_frame_type.grid(row=1, column=1)
 
@@ -73,24 +77,24 @@ class GenerateBarcodeWindow():
         self.color_metric_var.set("Average")
 
         # Dropdown menu for the color metric selection
-        dropdown_color_metric = tkinter.OptionMenu(self.window, self.color_metric_var, "Average", "Median", "Mode",
+        dropdown_color_metric = ttk.OptionMenu(self.window, self.color_metric_var, "Average", "Median", "Mode",
                                                    "Top-dominant", "Weighted-dominant", "Bright", "Brightest")
         dropdown_color_metric.grid(row=2, column=1)
 
         # Label prompt for the skip over specification
-        self.skip_over_label = tkinter.Label(self.window, text="Start at (frames): ")
+        self.skip_over_label = ttk.Label(self.window, text="Start at (frames): ")
         self.skip_over_label.grid(row=0, column=2)
 
         # Label prompt for the sample rate specification
[…]
