+++ setup.py
@@ -17,9 +17,9 @@ setup(name='kalmus',
       url='https://github.com/KALMUS-Color-Toolkit/KALMUS',
       author='Yida Chen, Eric Faden, Nathan Ryan',
       author_email='yc015@bucknell.edu',
-      license='MIT',
+      license='GPLv3',
       classifiers=[
-          "License :: OSI Approved :: MIT License",
+          "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
           "Programming Language :: Python :: 3",
           "Programming Language :: Python :: 3.7",
           "Programming Language :: Python :: 3.8",
