### 2021-03-11 ff891cd
update tutorial for command line ui
