+++ kalmus/barcodes/Barcode.py
@@ -25,8 +25,10 @@ def foreback_segmentation(frame):
     Segmented the input frame into two parts: foreground and background, using the GrabCut
 
     :param frame: Input frame
+    :type frame: numpy.ndarray
     :return: 1D image of the foreground part of the image, and 1D image of the background part of the image \
              Expected shape== Number of pixels x channels
+    :rtype: numpy.ndarray, numpy.ndarray
     """
     fore_frame, back_frame = Artist.grabcut_foreback_segmentation(frame, start_row=0, row_size=frame.shape[0] - 1,
                                                                   start_col=frame.shape[1] // 6,
@@ -36,20 +38,26 @@ def foreback_segmentation(frame):
 
 class Barcode:
     """
-    Barcode Class
+    Barcode Class. Base class for ColorBarcode and BrightnessBarcode
+
+    :param color_metric: The metric for computing the color of the frame
+    :type color_metric: str
+    :param frame_type: The type of frame sampling
+    :type frame_type: str
+    :param sampled_frame_rate: Frame sample rate: the frame sampled from every sampled_frame_rate.
+    :type sampled_frame_rate: int
+    :param skip_over: The number of frames to skip with at the beginning of the video
+    :type skip_over: int
+    :param total_frames: The total number of frames (computed) included in the barcode
+    :type total_frames: int
+    :param barcode_type: The type of the barcode
+    :type barcode_type: str
     """
 
     def __init__(self, color_metric, frame_type, sampled_frame_rate=1, skip_over=0, total_frames=10,
                  barcode_type=None):
         """
         Initialize the barcode with the given parameters
-
-        :param color_metric: The metric for computing the color of the frame
-        :param frame_type: The type of frame sampling
-        :param sampled_frame_rate: Frame sample rate: the frame sampled from every sampled_frame_rate.
-        :param skip_over: The number of frames to skip with at the beginning of the video
-        :param total_frames: The total number of frames (computed) included in the barcode
-        :param barcode_type: The type of the barcode
         """
         self.color_metric = color_metric
         self.frame_type = frame_type
@@ -87,6 +95,7 @@ class Barcode:
         Read in the video from the given path
 
         :param video_path_name: The path to the video file
+        :type video_path_name: str
         """
         self.video = cv2.VideoCapture(video_path_name)
 
@@ -111,6 +120,7 @@ class Barcode:
         Function run the get_letter_box_from_frames helper function by num_sample times and take the median of bounds
 
         :param num_sample: Number of times running the get_letter_box_from_frames
+        :type num_sample: int
         """
         possible_indexes = np.arange(self.film_length_in_frames // 6, self.film_length_in_frames * 5 // 6, 1)
         frame_indexes = np.random.choice(possible_indexes, num_sample, replace=True)
@@ -140,7 +150,9 @@ class Barcode:
         Remove the letter box from the frame using the known letter box bounds
 
         :param frame: Input original frame with letter box
+        :type frame: numpy.ndarray
         :return: Cropped frame without letter box
+        :rtype: numpy.ndarray
         """
         frame = frame[self.low_bound_ver: self.high_bound_ver, self.low_bound_hor: self.high_bound_hor, ...]
         return frame
@@ -150,7 +162,9 @@ class Barcode:
         Process the original frame by cropping out the letter box and resample frame using the given frame type
 
         :param frame: Input orignal frame
+        :type frame: numpy.ndarray
         :return: The processed and sampled frame
+        :rtype: numpy.ndarray
         """
         frame = self.remove_letter_box_from_frame(frame)
         if self.rescale_frames_in_generation:
@@ -182,6 +196,14 @@ class Barcode:
         return processed_frame
 
     def _resize_frame(self, frame):
+        """
+        resize the input frame with a factor of self.rescale_frame_factor
+
[…]
