+++ docs/_sources/barcodes/BrightnessBarcode.rst.txt
@@ -1,0 +1,8 @@
+BrightnessBarcode Class
+**************************
+
+.. automodule:: kalmus.barcodes.Barcode
+    :members: BrightnessBarcode
+    :undoc-members:
+    :show-inheritance:
+    :noindex:
