### 2021-03-13 2ff831f
update the guide for advanced usage of KALMUS's API
