### 2021-03-13 9e4d73f
update the guide for advanced usage of KALMUS's API
