### 2021-03-14 8c535d3
update guide for CLI
