### 2021-03-14 34c816f
update the argument table on user guide for CLI
