### 2021-03-14 def657b
update the requirements.txt file
