+++ requirements.txt
@@ -18,7 +18,6 @@ joblib==1.0.0
 jsonschema==3.0.1
 jupyter==1.0.0
 jupyter-console==6.0.0
-kalmus==1.3.6
 kiwisolver==1.3.1
 MarkupSafe==1.1.1
 matplotlib==3.2.2
@@ -45,7 +44,7 @@ pytz==2019.1
 PyWavelets==1.1.1
 pywinpty==0.5.5
 qtconsole==4.5.2
-scikit-image==0.16.2
+scikit-image>=0.16.2
 scikit-learn==0.24.0
 scipy==1.5.4
 Send2Trash==1.5.0
