### 2021-03-18 491dfd0
update for the security alert of Pillow requirement
