+++ requirements.txt
@@ -13,7 +13,7 @@ ipython==7.7.0
 ipython-genutils==0.2.0
 ipywidgets==7.5.0
 jedi==0.14.1
-Jinja2==2.10.1
+Jinja2>=2.11.3
 joblib==1.0.0
 jsonschema==3.0.1
 jupyter==1.0.0
