+++ PyPI_README.md
@@ -32,5 +32,5 @@ The KALMUS API reference is now available on
 
 # Update Log
 
-The full update log (from [v1.3.0](https://pypi.org/project/kalmus/1.3.0/) to [v1.3.7](https://pypi.org/project/kalmus/)) 
+The full update log (from [v1.3.0](https://pypi.org/project/kalmus/1.3.0/) to [v1.3.8](https://pypi.org/project/kalmus/)) 
 is now available on [https://kalmus-color-toolkit.github.io/KALMUS/update_log.html](https://kalmus-color-toolkit.github.io/KALMUS/update_log.html)
