+++ kalmus/utils/artist.py
@@ -18,17 +18,23 @@ def compute_dominant_color(image, n_clusters=3, max_iter=10, threshold_error=1.0
 
     :param image: input image. Either a multi-channel color image or a grayscale image (2D image) \
     Expected shape of image is height x width x (channels)
+    :type image: numpy.ndarray
     :param n_clusters: number of clusters
+    :type n_clusters: int
     :param max_iter: maximum iterations before terminating kmeans clustering
+    :type max_iter: int
     :param threshold_error: threshold error for terminating kmeans clustering. kmean clustering terminate \
     when the errors between the current computed and previous computed cluster center is under this \
     threshold
+    :type threshold_error: float
     :param attempts: Number of attempts to rerun the Kmeans clustering with different initial cluster \
     centers. Since kmeans clustering randomly choose n number of cluster in its initialization process.
+    :type attempts: int
     :return: An array of n cluster centers and an array of relative size (in percentage) of the clusters. \
     Expected output shape: n_clusters x channels (cluster centers), n_clusters (relative size in percentage) \
     E.g. for an input color image with 3 channels and n_clusters = 3  Output can be [[255, 255, 255], \
     [126, 75, 198], [186, 145, 122]], [0.4, 0.5, 0.1]
+    :rtype: numpy.ndarray, numpy.ndarray
     """
     assert len(image.shape) >= 2, "The input must be a 2 dimensional image"
     # Flatten the image
@@ -50,7 +56,9 @@ def flatten_image(image):
     with shape==[height x width, channels]
 
     :param image: Input 2D image (either multi-channel color image or greyscale image)
+    :type image: numpy.ndarray
     :return: The flatten 1D image. shape==(height x width, channels)
+    :rtype: numpy.ndarry
     """
     assert len(image.shape) >= 2, "The input image must be a 2 Dimensional image"
     if len(image.shape) == 3:
@@ -65,10 +73,12 @@ def compute_percents_of_labels(label):
     Compute the ratio/percentage size of the labels in an labeled image
 
     :param label: the labeled 2D image
+    :type label: numpy.ndarray
     :return: An array of relative size of the labels in the image. Indices of the sizes in the array \
     is corresponding to the labels in the labeled image. E.g. output [0.2, 0.5, 0.3] means label 0's size \
     is 0.2 of the labeled image, label 1' size is 0.5 of the labeled image, and label 2's size is 0.3 of \
     the labeled image.
+    :rtype: numpy.ndarray
     """
     # Get the bins of the histogram. Since the last bin of the histogram is [label, label+1]
     # We add 1 to the number of different labels in the labeled image when generating bins
@@ -88,11 +98,14 @@ def compute_mode_color(image, bin_size=10):
 
     :param image: either a multi-channel color image or a single channel greyscale image \
     Expected shape of image: height x width x (channels)
+    :type image: numpy.ndarray
     :param bin_size: Histogramize the input image. Color/intensity of each pixel in the \
     image will be accumulate in the bins with size==bin_size. The output mode color/intensity \
     is always an integer multiple of the bin_size.
+    :type bin_size: int
     :return: The mode color of the image (modes of the channels), shape=channels, and counts of the \
     mode colors happened in the input image, shape==channels
+    :rtype: numpy.ndarray, numpy.ndarray
     """
     image = flatten_image(image)
     image = image // bin_size
@@ -105,7 +118,9 @@ def compute_mean_color(image):
     Compute the average/mean color of the input multi-channel image or greyscale image.
 
     :param image: input image. Either a multi-channel color image or a single channel greyscale image
+    :type image: numpy.ndarray
[…]
