+++ docs/_sources/utils/Artist.rst.txt
@@ -1,7 +1,0 @@
-Artist Utility
-*****************
-
-.. automodule:: kalmus.utils.artist
-    :members: compute_mean_color, compute_median_color, compute_mode_color, compute_dominant_color, compute_percents_of_labels, flatten_image, compute_brightest_color_and_brightness, find_bright_spots, random_sample_pixels, grabcut_foreback_segmentation, watershed_segmentation, color_of_regions, get_contrast_matrix_and_labeled_image, contrast_between_regions, contrast_ratio, get_rag, rag_to_matrix, find_letter_box_from_videos, get_letter_box_from_frames
-    :undoc-members:
-    :show-inheritance:
