+++ kalmus/utils/visualization_utils.py
@@ -11,15 +11,24 @@ def show_color(color, figure_size=(9, 6), title="",
     Plot the rgb color.
 
     :param color: 1D array contains the R, G, and B channel values
+    :type color: numpy.ndarray
     :param figure_size: Size of the plotted figure
+    :type figure_size: tuple
     :param title: The title of the plotted color
+    :type title: str
     :param axis_off: True to set the axis of the plot figure off. False to have the axis on the side of \
     figure.
+    :type axis_off: bool
     :param save_image: True to save the plot figure into the path that user provide. False not to save \
     the plot figure
+    :type save_image: bool
     :param file_name: The filename of the saved figure. The default path to the saved image is test.png
-    :param return_color: Return the numpy array (shape==30 x 30 x 3) of color block if True. Directly plot the color if False
+    :type file_name: str
+    :param return_color: Return the numpy array (shape==30 x 30 x 3) of color block if True. Directly plot the color \
+    if False
+    :type return_color: bool
     :return: Return the numpy array of color block if return_color is True.
+    :rtype: numpy.ndarray
     """
     # Generate a block image where each pixel of this block image is the plotted RGB color
     clr = np.ones(shape=(30, 30, 3), dtype=np.uint8) * np.uint8(color)
@@ -47,15 +56,25 @@ def show_colors_in_sequence(colors, figure_size=(9, 6), title="", axis_off=False
     Plot a sequence of RGB colors either horizontally or vertically in line
 
     :param colors: 2D array of RGB colors. Expected sequence shape==Number of colors x channels (3)
+    :type colors: numpy.ndarray
     :param figure_size: the size of the plot figure
+    :type figure_size: tuple
     :param title: the title of the plot
+    :type title: str
     :param axis_off: True to set the axis of the figure off. False to have the axis with the figure
+    :type axis_off: bool
     :param save_image: True to save the plot figure. False do not save figure
+    :type save_image: bool
     :param file_name: the path of the saved figure
+    :type file_name: str
     :param horizontal: True to plot the sequence of colors horizontally, False to plot the sequence of colors \
                        vertically
+    :type horizontal: bool
     :param return_color_sequence: Return the numpy array (shape==30 x (30 * num_colors) x 3) of color sequence if True \
                                   Directly plot the color sequence if False
+    :type return_color_sequence: bool
+    :return: Return the numpy array of the color sequence image if return_color_sequence is True.
+    :rtype: numpy.ndarray
     """
     # A chain/sequence of color blocks
     clr_chain = []
@@ -96,16 +115,27 @@ def show_color_matrix(color_2d_array, mode="truncate", figure_size=(9, 4),
 
     :param color_2d_array: the input color matrix. Expected the shape of the color matrix to be row x col x channels \
     channels should 3 (R, G, and B channels of RGB colors)
+    :type color_2d_array: numpy.ndarray
     :param mode: mode for displaying color matrix. "truncate" or "padding"
+    :type mode: str
     :param figure_size: the size of the figure
+    :type figure_size: tuple
     :param title: the title of the plot
+    :type title: str
     :param axis_off: True to set the axis of the figure off
+    :type axis_off: bool
     :param save_image: True to save the plot figure, False do not save the figure
+    :type save_image: bool
     :param file_name: the path to the saved figure
+    :type file_name: str
     :param return_matrix: True to return the processed color matrix back False not to return anything
+    :type return_matrix: bool
     :param return_figure: Return the plotted figure, if true. Directly plot the color matrix if false
+    :type return_figure: bool
     :return: the processed color matrix. Depending on the display mode, the return color matrix will have padding \
-    white colors if mode is "padding", or the last row of color matrix will be truncated if mode is "truncate"
[…]
