### 2021-03-25 b04dcbb
update the doc for measure utils
