### 2021-03-26 #2980 @gkthiruvathukal
@yc015 Thank you. Even things got insanely busy over there. Apologies for the delay in following up.

@elektrobohemian or @macrocosme, just wondering whether either of you can help review this JOSS submission? I think I have asked macrocosme recently but perhaps your workload has eased up a bit now!

Please let me know so I can get review started.
