### 2021-03-30 #2980 @macrocosme
Thank you for the suggestion @gkthiruvathukal. While it seems like an interesting project, I unfortunately have no time available for this task at the moment.
