+++ requirements.txt
@@ -36,7 +36,7 @@ pickleshare==0.7.5
 Pillow>=8.1.1
 prometheus-client==0.7.1
 prompt-toolkit==2.0.9
-Pygments==2.4.2
+Pygments>=2.7.4
 pyparsing==2.4.1.1
 pyrsistent==0.15.4
 python-dateutil==2.8.0
