### 2021-04-08 #3156 @whedon
[REVIEW]: KALMUS: tools for color analysis of films
**Submitting author:** @yc015 (<a href="http://orcid.org/0000-0002-4018-6095">Yida Chen</a>)
**Repository:** <a href="https://github.com/KALMUS-Color-Toolkit/KALMUS" target ="_blank">https://github.com/KALMUS-Color-Toolkit/KALMUS</a>
**Version:** v1.3.9
