### 2021-04-12 4fd7e01
correct rtype in sphinx-doc
