### 2021-04-15 da9a76f
update the text in Markdown tutorial for CLI
