+++ notebooks/USAGE_COMMAND_LINE_UI.md
@@ -19,9 +19,9 @@ In this Markdown Tutorial I will introduce:
 # Notice
 
 The Command-line Interface of KALMUS is used specifically for **generating barcode** and **saving** it into 
-reloadable (via both [API](user_guide_for_kalmus_api.ipynb) and [GUI](user_guide_for_kalmus_gui.ipynb)) JSON object.
+reloadable (via both [API](user_guide_for_kalmus_api.ipynb) and [GUI](user_guide_for_kalmus_gui.ipynb)) JSON objects.
 
-We recommend our users to first walk through the IPython notebook Tutorials on either 
+We recommend our users first walk through the IPython notebook Tutorials on either 
 [KALMUS's Graphic user interface](user_guide_for_kalmus_gui.ipynb) or [KALMUS's Application programming interface](user_guide_for_kalmus_api.ipynb), 
 in which the full functionality of KALMUS is covered.
 
@@ -30,15 +30,15 @@ through this **Command-line interface**.
 
 # Installation Guide
 
-There are two ways that you could install KALMUS on you local machine:  
+There are two ways that you could install KALMUS on your local machine:  
 1. (**Recommended**) Get the latest distribution of KALMUS from PyPI ([KALMUS Project Page on PyPI](https://pypi.org/project/kalmus/)).  
-Use command `$ pip install kalmus` or `$ pip install --upgrade kalmus` (if kalmus has been installed) to install the latest version of KALMUS package. All dependencies should be automatically installed during this process.
+Use command `$ pip install kalmus` or `$ pip install --upgrade kalmus` (if kalmus has been installed) to install the latest version of the KALMUS package. All dependencies should be automatically installed during this process.
 
-2. Alternatively, you could install the KALMUS locally by first cloning the GitHub repo of Kalmus ([GitHub page](https://github.com/KALMUS-Color-Toolkit/KALMUS)). Then, move to the top directory of cloned KALMUS project and install using command `pip install .` 
+2. Alternatively, you could install the KALMUS locally by first cloning the GitHub repo of Kalmus ([GitHub page](https://github.com/KALMUS-Color-Toolkit/KALMUS)). Then, move to the top directory of cloned KALMUS project and install using the command `pip install .` 
 
 **See our [Installation Guide](https://kalmus-color-toolkit.github.io/KALMUS/install.html) for more details.**
 
-Once the package is installed, you could verify the version of KALMUS package using command `$ pip show kalmus`  
+Once the package is installed, you could verify the version of the KALMUS package using the command `$ pip show kalmus`  
 ![KALMUS version](notebook_figures/kalmus_version.png)
 
 Alternatively, in version 1.3.7 and above, you can check the version of installed kalmus using its 
@@ -52,7 +52,7 @@ Alternatively, in version 1.3.7 and above, you can check the version of installe
 
 ## Important!
 
-The **Command-line interface** is a new feature added into the KALMUS in its 1.3.7 version. To use Command-line 
+The **Command-line interface** is a new feature added into the KALMUS in its 1.3.7 version. To use the Command-line 
 interface, you have to make sure the version of installed KALMUS is 1.3.7 or onward.
 
 # How do I use the command-line interface?
@@ -64,7 +64,7 @@ Notice that if you have installed
 kalmus>=1.3.7 but failed to start the kalmus-generator through this command, you could find the corresponding executable 
 in your_python_directory/Scripts/ directory on Windows or .local/bin/ directory on Linux.
 
-Check the available arguments for `kalmus-generator` command using the flag `-h`:
+Check the available arguments for the `kalmus-generator` command using the flag `-h`:
 
 ```
 Windows OS
@@ -90,9 +90,9 @@ Only the arguments marked with **(required)** must be specified in the kalmus-ge
[…]
