### 2021-04-16 b9005fc
Dependecy issue on scikit-image should be fixed. Test package on python 3.9
