### 2021-04-16 8cc801f
update a markdown version of GUI user guide for non-technical audience
