+++ notebooks/USAGE_GRAPHIC_USER_INTERFACE.md
@@ -1,0 +1,302 @@
+# Welcome to the Markdown User Guide for KALMUS (GUI)!
+
+In this tutorial, I will introduce:  
+1. **Installation of KALMUS package**
+2. **What is KALMUS for**
+    - Extract color information from film frames or brightness information from monochrome film frames using different color metrics and image sampling methods.   
+    - Generate a barcode representation of the color/brightness information of a film. 3D information in 2D representation. 
+    - Compare different barcodes globally through similarity measures on images. Interpret the difference through similarity scores.
+    - Compare segments of barcodes locally using functions embedded in GUI. Interpret the difference using domain knowledge and contextual information extracted by KALMUS.  
+3. **How to interact with KALMUS using its Graphic user interface**
+    - Visualize barcodes
+    - Generate barcodes
+    - Change barcodes
+    - Save barcodes
+    - Load barcodes
+    - Compare barcodes
+    
+## 1. Installation
+There are two ways that you could install KALMUS on your local machine:  
+1. (**Recommended**) Get the latest distribution of KALMUS from PyPI ([KALMUS Project Page on PyPI](https://pypi.org/project/kalmus/)).  
+Use command `$ pip install kalmus` or `$ pip install --upgrade kalmus` (if kalmus has been installed) to install the latest version of the KALMUS package. All dependencies should be automatically installed during this process.
+
+2. Alternatively, you could install the KALMUS locally by first cloning the GitHub repo of Kalmus ([GitHub page](https://github.com/KALMUS-Color-Toolkit/KALMUS)). Then, move to the top directory of cloned KALMUS project and install using the command `pip install .` 
+
+**See our [Installation Guide](https://kalmus-color-toolkit.github.io/KALMUS/install.html) for more details.**
+
+Once the package is installed, you could verify the version of KALMUS package using the command `$ pip show kalmus`  
+<img src="notebook_figures/kalmus_version.png" alt="drawing" width="800 px"/>
+
+## 2. What is KALMUS for?
+KALMUS is a Python package for the computational analysis of colors in films. It addresses how to best describe a film's color. This package is optimized for two purposes: **(1) various ways to measure, calculate and compare a film's color and (2) various ways to visualize a film's color.**
+
+KALMUS utilizes the movie barcode as a visualization of the film's color. It has a modularized pipeline for the generation of barcodes using different measures of color and region of interest in each film frame. It also provides a set of measures that allow users to compare different films' colors directly through this visualization.
+
+### 2.1 Barcode Generation
+
+Barcode supports __7 color metrics__ that measure the color of a frame and __5 frame types__ that specify which part of the frame will be used in the color measures.
+
+Below is a table of available combinations of color metric and frame type in barcode generation.  
+
+| frame_type \ color_metric | Average | Median |  Mode  | Top-dominant | Weighted-dominant | Brightest | Bright |
+| --------------------------| :-----: | :----: | :----: | :----------: | :---------------: | :-------: | :----: |
+| **Whole_frame**               | &#9745; |   &#9745;  |  &#9745; |      &#9745;     |        &#9745;    |    &#9745;    |   &#9745;   |
+| **High_contrast_region**      | &#9745; |   &#9745;  |  &#9745; |      &#9745;     |      &#9745;      |    &#9745;    |   &#x2612;  |
+| **Low_contrast_region**       | &#9745; |   &#9745;  |  &#9745; |      &#9745;     |      &#9745;      |    &#9745;    |   &#x2612;  |
[…]
