### 2021-04-16 a3704a5
add separation line between the sections in markdown tutorial for gui
