### 2021-04-18 KALMUS-Color-Toolkit/KALMUS#3 @thomakra
Suggestions on JOSS paper
Regarding the JOSS paper under review, here are some suggestions to improve the submission:

Please have a look at the references in the paper and make sure it is possible to find them online without too much effort by providing DOI's, URLs, etc.
When searching, I could not find anything resembling the Adams (2021) reference. Is it under review somewhere so that it is not available?

I notice that the main barcodes do not have axis labels. If I understand correctly, each pixel represents some statistic of the color of a single frame and these statistics are then ordered top-to-bottom, then left-to-right to form an image. I'm wondering if it would make it easier to get a feel for the length of a sequence if the x-axis ticks showed the total elapsed frames/seconds and the y-axis showed the frames/seconds per column with appropriate labels.

I think it would be nice to elaborate a little how the barcodes are generated. There is a very nice table [here](https://github.com/KALMUS-Color-Toolkit/KALMUS/blob/master/notebooks/USAGE_COMMAND_LINE_UI.md) giving an overview of the provided metrics and the types of region from which the metric is computed. Perhaps it would be nice to include it in the paper with a few lines describing it?

Other than that I think the paper looks nice and the package is well documented!
