### 2021-04-18 #3156 @thomakra
@gkthiruvathukal I'm mostly done with the review and have very few comments.
The paper is well written, the package is well documented and I was able to install and use the software as described.
I've added some suggestions in the [KALMUS repository](https://github.com/KALMUS-Color-Toolkit/KALMUS/issues/3).
Once the reference list (last remaining checkbox) has been improved to make the references easier to find online, I'm happy to recommend it for publication.
