+++ paper/joss-refs.bib
@@ -13,7 +13,8 @@
   number={1},
   pages={26--41},
   year={2008},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  note={\url{https://doi.org/10.1016/j.media.2007.06.004}}
 }
 
 
@@ -22,7 +23,8 @@ author = {Flueckiger, Barbara and Halter, Gaudenz},
 title = {Methods and Advanced Tools for the Analysis of Film Colors in Digital Humanities},
 journal = {Digital Humanities Quarterly: Special Issue Digital Humanities&Film Studies. Analyzing the Modalities of Moving Images}, 
 volume={14},
-number ={4}, 
+number ={4} 
+
 }
 
 @phdthesis{stutz,
@@ -48,7 +50,8 @@ number ={4},
   number={2},
   pages={71--94},
   year={2017},
-  publisher={JSTOR}
+  publisher={JSTOR},
+  note={\url{https://doi.org/10.5749/movingimage.17.2.0071}}
 }
 
 
@@ -67,7 +70,8 @@ number ={4},
   number={3},
   pages={443--453},
   year={1970},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  note={\url{https://doi.org/10.1016/b978-0-12-131200-8.50031-9}}
 }
 
 
@@ -79,7 +83,8 @@ number ={4},
   volume={5},
   number={45},
   pages={1800},
-  year={2020}
+  year={2020},
+  note={\url{https://doi.org/10.21105/joss.01800}}
 }
 
 @music{olsen,
@@ -101,7 +106,8 @@ number ={4},
   number={3},
   pages={119--129},
   year={2019},
-  organization={Wiley Online Library}
+  organization={Wiley Online Library},
+  note={\url{https://doi.org/10.1111/cgf.13676}}
 }
 
 
@@ -112,7 +118,8 @@ number ={4},
   volume={11},
   number={2},
   pages={206--223},
-  year={2002}
+  year={2002},
+  note={\url{https://doi.org/10.1117/1.1455011}}
 }
 
 
@@ -124,7 +131,8 @@ number ={4},
   number={1},
   pages={195--197},
   year={1981},
-  publisher={Elsevier Science}
+  publisher={Elsevier Science},
+  note={\url{https://doi.org/10.1016/0022-2836(81)90087-5}}
 }
 
 @unpublished{adams,
@@ -142,6 +150,7 @@ number ={4},
   number={4},
   pages={600--612},
   year={2004},
-  publisher={IEEE}
+  publisher={IEEE},
+  note={\url{https://doi.org/10.1109/tip.2003.819861}}
 }
 
