### 2021-04-18 KALMUS-Color-Toolkit/KALMUS#3 @yc015
Hi @thomakra, thank you so much for your comments. We are updating the missing DOIs of the citation. For the Adams (2021) reference, the paper is still in progress and not being published yet. Therefore, we don't have a DOI corresponding to that article. Do you have any recommendations about addressing the Adams (2021) reference?
