+++ paper/joss-refs.bib
@@ -14,7 +14,7 @@
   pages={26--41},
   year={2008},
   publisher={Elsevier},
-  note={\url{https://doi.org/10.1016/j.media.2007.06.004}}
+  url = {https://doi.org/10.1016/j.media.2007.06.004}
 }
 
 
@@ -51,7 +51,7 @@ number ={4}
   pages={71--94},
   year={2017},
   publisher={JSTOR},
-  note={\url{https://doi.org/10.5749/movingimage.17.2.0071}}
+  url = {https://doi.org/10.5749/movingimage.17.2.0071}
 }
 
 
@@ -71,7 +71,7 @@ number ={4}
   pages={443--453},
   year={1970},
   publisher={Elsevier},
-  note={\url{https://doi.org/10.1016/b978-0-12-131200-8.50031-9}}
+  url = {https://doi.org/10.1016/b978-0-12-131200-8.50031-9}
 }
 
 
@@ -107,7 +107,7 @@ number ={4}
   pages={119--129},
   year={2019},
   organization={Wiley Online Library},
-  note={\url{https://doi.org/10.1111/cgf.13676}}
+  url = {https://doi.org/10.1111/cgf.13676}
 }
 
 
@@ -119,7 +119,7 @@ number ={4}
   number={2},
   pages={206--223},
   year={2002},
-  note={\url{https://doi.org/10.1117/1.1455011}}
+  url = {https://doi.org/10.1117/1.1455011}
 }
 
 
@@ -132,7 +132,7 @@ number ={4}
   pages={195--197},
   year={1981},
   publisher={Elsevier Science},
-  note={\url{https://doi.org/10.1016/0022-2836(81)90087-5}}
+  url = {https://doi.org/10.1016/0022-2836(81)90087-5}
 }
 
 @unpublished{adams,
@@ -151,6 +151,6 @@ number ={4}
   pages={600--612},
   year={2004},
   publisher={IEEE},
-  note={\url{https://doi.org/10.1109/tip.2003.819861}}
+  url={https://doi.org/10.1109/tip.2003.819861}
 }
 
