+++ paper/joss-refs.bib
@@ -84,7 +84,7 @@ number ={4}
   number={45},
   pages={1800},
   year={2020},
-  note={\url{https://doi.org/10.21105/joss.01800}}
+  url = {https://doi.org/10.21105/joss.01800}
 }
 
 @music{olsen,
