+++ paper/joss-refs.bib
@@ -62,6 +62,7 @@ number ={4}
   author={Yumibe, Joshua},
   year={2012},
   publisher={Rutgers University Press}
+  doi={10.1080/17460654.2013.837593}
 }
 
 @article{needleman,
@@ -88,7 +89,7 @@ number ={4}
   pages={1800},
   year={2020},
   url = {https://doi.org/10.21105/joss.01800},
-  doi = {https://doi.org/10.21105/joss.01800}
+  doi = {10.21105/joss.01800}
 }
 
 @music{olsen,
