+++ paper/joss-refs.bib
@@ -61,7 +61,8 @@ number ={4}
   title={Moving color: Early film, mass culture, modernism},
   author={Yumibe, Joshua},
   year={2012},
-  publisher={Rutgers University Press}
+  publisher={Rutgers University Press},
+  url={https://doi.org/10.1080/17460654.2013.837593},
   doi={10.1080/17460654.2013.837593}
 }
 
