### 2021-04-18 342b623
using the temporal axis in the GUI display
