+++ kalmus/barcodes/BarcodeGenerator.py
@@ -65,6 +65,10 @@ def build_barcode_from_json(path_to_json, barcode_type="Color"):
     if "fps" in object_dict.keys():
         if object_dict["fps"] is not None:
             barcode.fps = float(object_dict["fps"])
+        else:
+            barcode.fps = 30
+    else:
+        barcode.fps = 30
 
     barcode.film_length_in_frames = int(object_dict["film_length_in_frames"])
 
