### 2021-04-18 KALMUS-Color-Toolkit/KALMUS#3 @yc015
Hello @thomakra, here are some updates on your suggestions!

(1) For the DOIs and URLs, all the missing DOIs have been added to the paper except for the Adams (2021), which is still in progress. We wish to hear your suggestion about how to address this in-progress article.

(2) For the x-axis ticks and y-axis label, it would help a lot in direct comparison using the temporal ticks! I have just changed the main barcodes following your suggestion. Here is how it looks like now:

![main_barcode](https://user-images.githubusercontent.com/69224866/115161580-90bab780-a06c-11eb-95b8-9101902f00a9.png)

The y-label shows the seconds per column rate, and the x-axis ticks show the total elapsed time.

(3) For the [table](https://github.com/KALMUS-Color-Toolkit/KALMUS/blob/master/notebooks/USAGE_COMMAND_LINE_UI.md), we have included it in the paper with some descriptions of available metrics and types of region. Here is the link to the new article proof: [https://github.com/openjournals/joss-papers/blob/joss.03156/joss.03156/10.21105.joss.03156.pdf](https://github.com/openjournals/joss-papers/blob/joss.03156/joss.03156/10.21105.joss.03156.pdf).

Thank you again for these invaluable suggestions! We are looking forward to hearing your feedback.
