### 2021-04-20 KALMUS-Color-Toolkit/KALMUS#3 @thomakra
Thank you for your changes. I think it looks very good!

> Hi @thomakra, thank you so much for your comments. We are updating the missing DOIs of the citation. For the Adams (2021) reference, the paper is still in progress and not being published yet. Therefore, we don't have a DOI corresponding to that article. Do you have any recommendations about addressing the Adams (2021) reference?

Regarding the Adams 2021 submission: I think I would check with the editor if they have any recommendations from the journal's side. Since the reference is used as an example of use, perhaps it can be omitted. If the point for including it is to link the software paper with the other paper, perhaps the JOSS paper can be referenced in the Adams paper instead? I guess the course of action may be dictated by which paper gets accepted first. At the very least it should state the journal it has been submitted to and say something like "(submitted)" or "(in review)".

One final thing that puzzled me: The review issue in GitHub references "Version: v.1.3.5", however I can't see any tags in your repository. Where do you have these version numbers?
