### 2021-04-20 KALMUS-Color-Toolkit/KALMUS#3 @yc015
Hello @thomakra, thank you for your suggestions on Adams reference! I agree it may be more helpful to reference this paper in the Adams paper instead, and we would like to wait for the editor's recommendations.

For the versioning, we archived our package through PyPI. The Version: v.1.3.5 is corresponding to the package's 1.3.5 release on PyPI: [link to v1.3.5 on PyPI](https://pypi.org/project/kalmus/1.3.5/).

After we submit the JOSS paper, the package has progressed to version: v1.3.8 in the past three months. [link to v1.3.8 on PyPI](https://pypi.org/project/kalmus/) Should we update the version number in our JOSS paper to reflect this change?
