### 2021-04-20 KALMUS-Color-Toolkit/KALMUS#3 @thomakra
Thanks for the clarification.
I guess I just expected the versions to be tags in the git repository.
No worries.

I have no further comments and I think you can safely close the issue and
wait for comments from the other reviewer.
