### 2021-04-26 KALMUS-Color-Toolkit/KALMUS#3 @elektrobohemian
in addition to the remarks made by @thomakra, I have some recommendations and a central question.

you are stating that color differences are subtle (p. 1) and that studies are also carried out on historical film releases (p. 3) which are most likely retro-digitized. thus, I was wondering why the whole paper is not addressing issues such as color spaces and gamut mapping. In case you deal with two retro-digitized film copies of the same movie your tool would create different color analyses just because of the fact that the digitizers used different devices and color profiles and decided to use different color spaces during processing -although the color difference is not present in the original material. (the same holds true for different editions with additional post-production runs or introduction of HDR or the like.) I would address this issue at least as a potential risk when using KALMUS. alternatively, you could add a warning to the tool itself or provide options to carry out gamut mappings if color profiles mismatch. the color profile issues of scanners etc. are definitely out of the scope of your paper. what has been digitized without care cannot be healed by your software, maybe you want to discuss this issue as you cannot control all the variables affecting color.

besides this conceptual issue, I have only minor hints on improving the readability of the text:

* (p. 2, l. 39) Be a little more precise why existing movie barcode software is inadequate.
* (p. 2, l. 42) "a barcode's data" is relatively vague, maybe you could give an example or refer to later parts of the paper.
* (p. 2, l. 44ff.) I would expand the list as a separate text block
* (p. 3, l. 53ff.) this could be a potential location to discuss the issue presented above
