### 2021-04-27 1fff8a4
Update the installation warning for Apple M1 processors users
