### 2021-04-27 a3d46b9
Clarify the warning for Apple M1 users
