+++ README.md
@@ -73,7 +73,7 @@ Alternatively, in version 1.3.7 and above, you can check the version of installe
 
 ## For users with Apple M1 Chip (arm64 Architecture)
 
-As @elektrobohemian mentioned in [issue #4](https://github.com/KALMUS-Color-Toolkit/KALMUS/issues/4), kalmus cannot build natively on Apple M1 processors because of kalmus's dependencies on scikit-image, scikit-learn, scipy, and numpy. You may be able to install kalmus under a Rosetta emulation with Python 3.7. 
+As @elektrobohemian mentioned in [issue #4](https://github.com/KALMUS-Color-Toolkit/KALMUS/issues/4), kalmus cannot build natively on Apple M1 processors because of kalmus's dependencies on NumPy. You may be able to install kalmus under a Rosetta emulation with Python 3.7. 
 
 # Get Started
 
