### 2021-04-27 a8c8442
redirect the warning in notebooks/README.md to the main README.md
