### 2021-04-27 0ee63a9
update the installation Guide in API documentation page
