### 2021-05-03 3d87028
update the joss paper for hints 1, 2, 3
response to reviewer's suggestion 1, 2, 3 in issue #3
