+++ paper/joss-paper.md
@@ -49,14 +49,14 @@ KALMUS is built around a visualization of films known as movie barcodes; it prod
 [^1]: In OpenCV [@opencv], for example, if a frame is decimated to a single pixel, its color is determined by area interpolation, a weighted average of the RGB-values of the pixels.
 
 KALMUS allows for the analysis of a film's color by:  
-\begin{block}
+::: {.block}
 1. Providing an interface that takes in either a video file or a JSON file (a sample JSON file can be found [here](https://github.com/KALMUS-Color-Toolkit/KALMUS/blob/master/kalmus/data/mission_impossible_Bright_Whole_frame_Color.json)).
 2. Allowing for the computation of a frame's color in a number of ways (dominant, median, etc).
 3. Allowing for the computation of a film's color in a number of ways. 
 4. Providing implementations of ways to compare colors of two films. 
 5. Providing implementations of ways to visualize the color of a film.
 6. Allowing the user to download color data as a CSV or JSON file.  See Figure 2 for an explanation of KALMUS's user interface.
-\end{block}
+:::  
 
 ![Image A shows barcodes for a single film in which the color of a frame was calculated in two ways:  the top one uses the frame's brightest color and the bottom one uses the frame's median color.  Image B shows the interface for generating a barcode: one can process a film in parallel, choose a sampling rate, and choose a method for determining a frame's color.  Image C is the result of selecting a pixel in a barcode to give a user an idea of where in the film they are.  Image D is a table of similarity metrics between the two barcodes.  Image E is an interactive 3D plot of the RGB values in one of the barcodes.](images-joss/kalmus-interface.jpg)
 
