+++ paper/joss-paper.md
@@ -42,25 +42,25 @@ Three reasons emerge: (1) film color remains subtle and almost subliminal by des
 
 Recently, several digital tools have emerged that analyze a film’s color.  The ability to capture film stills from DVD or streaming sources and analyze the frame’s palette in off-the-shelf tools like [Adobe Color](http://color.adobe.com) or Adobe Photoshop provide insights into individual frames.  Additionally, many open-source projects provide implementations of state-of-the-art computer vision algorithms applied to moving images.  Some, such as the Distance Viewing Toolkit [@arnold], are very general and handle several aspects of a moving image at once (sound, color, camera angle, etc.).  Others, such as VIAN [@vian], allow for a "closer", more interactive distant viewing.  
 
-KALMUS is built around a visualization of films known as movie barcodes; it produces an overall color palette “snapshot” of a film.  Artists like [Jeffrey Moser](http://www.jeffreymoser.com/) or programmers such as [Charlie Clark](https://thecolorsofmotion.com/about), generate movie barcodes by reducing each film frame to a single color[^1] and then stitching these colors into a mosaic.  Existing movie barcode software, without relevant statistical tools, is inadequate for studying films.  While they provide a fascinating visual “overview” of a film’s color palette, the software focuses more on visualization rather than providing meaningful analytical data.  KALMUS generates barcodes but also generates a quantitative color palette (see Figure 2) so that users can carry out their own analyses.  See Figure 1 for examples of movie barcodes.  
+KALMUS is built around a visualization of films known as movie barcodes; it produces an overall color palette “snapshot” of a film.  Artists like [Jeffrey Moser](http://www.jeffreymoser.com/) or programmers such as [Charlie Clark](https://thecolorsofmotion.com/about), generate movie barcodes by reducing each film frame to a single color[^1] and then stitching these colors into a mosaic.  Existing movie barcode software, without relevant statistical tools, is inadequate for studying films.  While they provide a fascinating visual “overview” of a film’s color palette, the software focuses more on visualization rather than providing meaningful analytical data.  KALMUS generates barcodes but also generates a quantitative color profile (see Figure 2) so that users can carry out their own analyses.  See Figure 1 for examples of movie barcodes.  
 
 ![The KALMUS interface.](images-joss/kalmus_figure1_1.jpg)
 
 [^1]: In OpenCV [@opencv], for example, if a frame is decimated to a single pixel, its color is determined by area interpolation, a weighted average of the RGB-values of the pixels.
 
 KALMUS allows for the analysis of a film's color by:  
-1. Providing an interface that takes in either a video file or a JSON file (a sample JSON file can be found [here](https://github.com/KALMUS-Color-Toolkit/KALMUS/blob/master/kalmus/data/mission_impossible_Bright_Whole_frame_Color.json)).
-2. Allowing for the computation of a frame's color in a number of ways (dominant, median, etc).
-3. Allowing for the computation of a film's color in a number of ways. 
-4. Providing implementations of ways to compare colors of two films. 
-5. Providing implementations of ways to visualize the color of a film.
-6. Allowing the user to download color data as a CSV or JSON file.  See Figure 2 for an explanation of KALMUS's user interface. 
+1. Providing an interface that takes in either a video file or a JSON file (a sample JSON file can be found [here](https://github.com/KALMUS-Color-Toolkit/KALMUS/blob/master/kalmus/data/mission_impossible_Bright_Whole_frame_Color.json)).  
+2. Allowing for the computation of a frame's color in a number of ways (dominant, median, etc).  
+3. Allowing for the computation of a film's color in a number of ways.  
+4. Providing implementations of ways to compare colors of two films.  
[…]
