+++ paper/joss-paper.md
@@ -63,7 +63,7 @@ KALMUS allows for the analysis of a film's color by:
 
 KALMUS allows users to understand a film’s color palette and compare that palette to other films.  See Table 1 for options on what parts of the frame a user might want to analyze and the various metrics a user can use to determine the color of a frame.  See Table 2 for information on how film colors can be compared.  Potentially, films from a certain time period or a particular genre might be aggregated to see if there are common color palettes[^2].  While film studies as a discipline has long depended on qualitative analysis, KALMUS provides a straightforward quantitative tool that can supplement the reading of a film.  For example, in [@adams] KALMUS is used to study color trends in Hollywood movies from 1990 to 2015.  In a class taught by the second author, students used KALMUS to explore how a film's color palette signaled narrative shifts and introduced significant characters.
 
-[^2]. We point out users should be aware of the fact that different color gamuts used in digitizing films may affect a film's original color palette.  KALMUS is designed to use the broadcast color gamut called REC 709.  
+[^2]: We point out users should be aware of the fact that different color gamuts used in digitizing films may affect a film's original color palette.  KALMUS is designed to use the broadcast color gamut called REC 709.  
 
 
 |                     | Mean	| Median   | Mode	| TD	| WD	| BR	| BP  |
