### 2021-05-04 KALMUS-Color-Toolkit/KALMUS#3 @yc015
Hello @elektrobohemian! Sorry for my late reply. Other authors and I have carefully discussed your remarkable question of Gamut mapping last week.

We added a footnote of warning for analyzing films that are digitized with different color spaces in p. 3, l. 65.

We also made the following changes to the text according to your hints:
1. (p. 2, l. 40) We answer why existing barcode software is inadequate.
2. (p. 2, l. 43ff.) We Illustrate the "barcode's data".
3. (p. 2, l. 46ff.) Expand the lists.

The newly compiled paper can be downloaded here: [https://raw.githubusercontent.com/openjournals/joss-papers/joss.03156/joss.03156/10.21105.joss.03156.pdf](https://raw.githubusercontent.com/openjournals/joss-papers/joss.03156/joss.03156/10.21105.joss.03156.pdf)

Thank you again for these remarkable suggestions! Please let us know if you have any following questions.
