### 2021-05-13 KALMUS-Color-Toolkit/KALMUS#3 @elektrobohemian
thanks for work. I have tested the software as well and couldn't find any issues. I recommend your interesting work for publication.
