### 2021-05-16 #3156 @yc015
@gkthiruvathukal Thank you! I have tagged the release and archived the software in Zenodo.
The tagged release is here [REL: v1.3.9 JOSS release of KALMUS](https://github.com/KALMUS-Color-Toolkit/KALMUS/releases/tag/v1.3.9)
The Zenodo DOI is 10.5281/zenodo.4765745  ([link to the Zenodo page](https://zenodo.org/record/4765745#.YKE4cqhKhPY)).
