### 2021-05-17 #3156 @kyleniemeyer
Hi @yc015, I'm the AEIC on duty this week doing some final checks on your submission before publishing.

In the list of references, I found some potential issues:
- Can you provide more detail on the Adams et al. reference? I can't seem to find it at all via a quick search.
- The Flueckiger and Halter ref should have 2020 as the year
