+++ paper/joss-refs.bib
@@ -23,6 +23,7 @@
 author = {Flueckiger, Barbara and Halter, Gaudenz},
 title = {Methods and Advanced Tools for the Analysis of Film Colors in Digital Humanities},
 journal = {Digital Humanities Quarterly: Special Issue Digital Humanities&Film Studies. Analyzing the Modalities of Moving Images}, 
+year={2020},
 volume={14},
 number ={4} 
 
