### 2021-05-17 #3156 @yc015
@kyleniemeyer Thank you so much for checking our submission!

> The Flueckiger and Halter ref should have 2020 as the year

The year of publication for this reference has been added.

> Can you provide more detail on the Adams et al. reference? I can't seem to find it at all via a quick search.

As we mentioned in the [issue #3](https://github.com/KALMUS-Color-Toolkit/KALMUS/issues/3#issuecomment-823007923), the Adams et al. paper is still in progress, but we are happy to omit it and its associated text from this submission. This reference is only used as an example of use for KALMUS.
