### 2021-05-17 #3156 @kyleniemeyer
@yc015 ah, I didn't see that issue/discussion. My/our preference would be to only include citations for things that are either already submitted, or somehow available openly via a preprint. Is that paper at the point of submitting a preprint?

However, if it isn't, then we can proceed as-is, but should add some note to the reference about it being "in preparation".
