+++ paper/joss-refs.bib
@@ -149,6 +149,7 @@ number ={4}
   author = {Adams, Meg and Chen, Yida and Faden, Eric and Ryan, Nathan C. and Zhang, Haohuan},
   title  = {The changing colors of {H}ollywood films},
   note   = {in preparation},
+  pubstate = {\bibstring{in preparation}},
   year   = {2021},
 }
 
