+++ paper/joss-refs.bib
@@ -145,11 +145,11 @@ number ={4}
   doi = {10.1016/0022-2836(81)90087-5}
 }
 
-@article{adams,
+@unpublished{adams,
   author = {Adams, Meg and Chen, Yida and Faden, Eric and Ryan, Nathan C. and Zhang, Haohuan},
   title  = {The changing colors of {H}ollywood films},
-  note   = {in preparation},
-  year   = {2021}
+  note   = "in preparation",
+  year   = {2021 in preparation}
 }
 
 @article{wang,
