### 2021-05-17 #3156 @yc015
@kyleniemeyer Thank you for your suggestion! We don't have a preprint for that paper yet. The pandemic slows down our progress on that project.

I added a note of "in preparation" to the Adams et al. reference.
