### 2021-05-18 #3156 @kyleniemeyer
@yc015 ah, could you move that out of the `year` field? that should just have the numeric year, since that is reflected in the in-text reference.

Also, in the text, could you change that specific citation command from `For example, in [@adams]` to `For example, in @adams`? That will result in "For example, in Adams et al. (2021)" rather than "For example, in (Adams et al., 2021)".
