+++ paper/joss-refs.bib
@@ -148,8 +148,8 @@ number ={4}
 @unpublished{adams,
   author = {Adams, Meg and Chen, Yida and Faden, Eric and Ryan, Nathan C. and Zhang, Haohuan},
   title  = {The changing colors of {H}ollywood films},
-  note   = "in preparation",
-  year   = {2021 in preparation}
+  notes   = "in preparation",
+  year   = {2021}
 }
 
 @article{wang,
