+++ paper/joss-refs.bib
@@ -148,9 +148,9 @@ number ={4}
 @unpublished{adams,
   author = {Adams, Meg and Chen, Yida and Faden, Eric and Ryan, Nathan C. and Zhang, Haohuan},
   title  = {The changing colors of {H}ollywood films},
-  note   = {in preparation},
-  pubstate = {in preparation},
-  year   = {2021}
+  % note   = {in preparation},
+  pubstate = {in preparation}
+  % year   = {2021}
 }
 
 @article{wang,
