### 2021-05-18 #3156 @yc015
@kyleniemeyer I have changed the command from `[@adams]` to `@adams`.

For the "in preparation" note, I moved it into the `pubstate` field. I removed the `year` field otherwise the `pubstate` field will not show in the reference if `year` exists.
